% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{add_loggers}
\alias{add_loggers}
\title{Add loggers to an object}
\usage{
add_loggers(x, loggers)
}
\arguments{
\item{x}{a node object}

\item{loggers}{a list of loggers}
}
\description{
Add loggers to an object
}
