% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{add_node}
\alias{add_node}
\title{Initialize a node and adds it to an rflow}
\usage{
add_node(obj, rflow, ..., conflict = "update", storage = NULL,
  connect = FALSE, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Initialize a node and adds it to an rflow
}
