% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{add_nodes}
\alias{add_nodes}
\title{batch init objects}
\usage{
add_nodes(objs, rflow, connect = TRUE, ..., verbose = TRUE)
}
\arguments{
\item{objs}{list of nodes' definitions}

\item{rflow}{an rflow object}

\item{connect}{logical;}

\item{verbose}{logical; print verbose output?}
}
\description{
batch init objects
}
