% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{add_prefix}
\alias{add_prefix}
\title{Print with prefix}
\usage{
add_prefix(x, prefix = "", color_main = NULL, color_prefix = NULL)
}
\arguments{
\item{x}{value to print}

\item{verbose_prefix}{prefix to be added after every new-line symbol}
}
\description{
Print with prefix
}
