% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{as_code}
\alias{as_code}
\title{Converts table of function information into lines of code}
\usage{
as_code(outputVar, operator, funs)
}
\description{
Converts table of function information into lines of code
}

