% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{as_r_expr}
\alias{as_r_expr}
\title{Returns R expression from either R expression or parsed R code}
\usage{
as_r_expr(r_code = NULL, r_expr = NULL)
}
\arguments{
\item{r_code}{character vector of R code; ignored when R expression is not NULL}

\item{r_expr}{optional R expression}
}
\description{
Returns R expression from either R expression or parsed R code
}
