% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\docType{data}
\name{cache}
\alias{cache}
\title{holds connection object}
\format{An object of class \code{environment} of length 0.}
\usage{
cache
}
\description{
holds connection object
}
\keyword{datasets}

