% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{cache_file}
\alias{cache_file}
\title{Get the file only cache engine.}
\usage{
cache_file(cache_dir)
}
\arguments{
\item{cache_dir}{Directory where to store the cache files.}
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
All cache data will be stored only on the local disk.
}
\examples{
\dontrun{
cache_in_file <- cache_file("path_to_cache_diretory")
}

}
\seealso{
Other cache functions: \code{\link{cache_memory_file}},
  \code{\link{cache_memory}}, \code{\link{default_cache}}
}
\concept{cache functions}
