% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{cache_memory}
\alias{cache_memory}
\title{Get the memory cache engine.}
\usage{
cache_memory()
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
Get the memory cache engine.
}
\seealso{
Other cache functions: \code{\link{default_cache}}
}
\concept{cache functions}
