% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{cache_memory_file}
\alias{cache_memory_file}
\title{Get the memory-file cache engine.}
\usage{
cache_memory_file()
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
The cache data will be stored both in memory (for quick access) and on
  disk (for persistence).
}
\seealso{
Other cache functions: \code{\link{cache_file}},
  \code{\link{cache_memory}}, \code{\link{default_cache}}
}
\concept{cache functions}
