% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_r_expr}
\alias{cat_r_expr}
\title{Pretty printing of R expressions}
\usage{
cat_r_expr(r_expr, verbose_prefix = "")
}
\arguments{
\item{verbose_prefix}{}
}
\description{
Pretty printing of R expressions
}
