% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{cat_with_prefix}
\alias{cat_with_prefix}
\title{Print with prefix}
\usage{
cat_with_prefix(x, prefix = "", sep = "\\n", ...)
}
\arguments{
\item{x}{value to print}

\item{verbose_prefix}{prefix to be added after every new-line symbol}
}
\description{
Print with prefix
}
