% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nodes.R
\name{check_fun}
\alias{check_fun}
\title{User-facing Function Checker}
\usage{
check_fun(fun_name, pkg, ns)
}
\arguments{
\item{fun_name}{Character scalar naming an object in a package's namespace}

\item{pkg}{Character scalar naming a package}

\item{ns}{Environment corresponding to package's namespace}
}
\value{
Logical scalar showing whether an object is a function with certain features
  or not
}
\description{
For a given object in some package, determine whether it is a function,
  it has at least one argument and has documentation.
}

