% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{clean_cache}
\alias{clean_cache}
\title{Delete/clean cache file/folder}
\usage{
clean_cache(x, ...)
}
\arguments{
\item{x}{a node or rflow object}

\item{...}{}
}
\description{
Delete/clean cache file/folder
}
