% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{clean_persistence}
\alias{clean_persistence}
\title{Delete/clean stored state file/folder}
\usage{
clean_persistence(x, ...)
}
\arguments{
\item{x}{a node or rflow object}

\item{...}{}
}
\description{
Delete/clean stored state file/folder
}
