% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{compose_id}
\alias{compose_id}
\title{Compose an id value from env and name}
\usage{
compose_id(x)
}
\arguments{
\item{obj}{object definition}
}
\value{
Node's id as a character value constructed from env + name fields.
}
\description{
Compose an id value from env and name
}
