% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{compute}
\alias{compute}
\alias{compute.R6Flow}
\alias{compute.Element}
\title{Trigger computation for an \code{R6Flow} or an \code{Element} object.}
\usage{
\method{compute}{R6Flow}(x, ...)

\method{compute}{Element}(x, ...)
}
\arguments{
\item{x}{A flow object, e.g. as returned by \code{\link{flow_fn}}.}

\item{...}{Any other arguments will be ignored.}
}
\value{
Logical, whether the result is available to be collected.
}
\description{
Trigger computation for an \code{R6Flow} or an \code{Element} object.
}
\details{
Unlike \code{collect}, it does not trigger an error if it fails 
   to compute and it does not return the actual result of the computation.
}
