% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{connect_nodes.rflow}
\alias{connect_nodes.rflow}
\title{batch connect all objects}
\usage{
\method{connect_nodes}{rflow}(rflow, ...)
}
\arguments{
\item{rflow}{an rflow object}
}
\description{
batch connect all objects
}
