% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{Convolutionary Neural Network Model}
\alias{ConvModel}
\alias{Convolutionary Neural Network Model}
\title{Convolutionary Neural Network Model}
\usage{
ConvModel(n_filters = 12, filter_shape = c(3, 3),
  activ_func = "logistic_regression", transform_method = "tf.expand_dims",
  pool_method = "tf.reduce_max", reduction_indices = c(1, 2),
  shape = c(-1, 12), ...)
}
\arguments{
\item{n_filters}{Number of filters}

\item{filter_shape}{A vector of integers representing the shape of filters}

\item{activ_func}{The activation function, e.g. logistic_regression, linear_regression}

\item{transform_method}{The transformation method used, a list of methods can be found on TensorFlow API doc}

\item{pool_method}{Pooling method. A list can be found on API doc}

\item{reduction_indices}{A vector of reduction indices used in pooling}

\item{shape}{The shape parameter passed to tf.reshape. See API doc.}

\item{...}{Additional parameters can be passed, such as strides, see a list here: https://github.com/google/skflow/blob/master/skflow/ops/conv_ops.py}
}
\description{
This function builds convolutionary neural network model with many tunable parameters
}

