% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{dbi}
\alias{dbi}
\title{Wrapper for DBI connection}
\usage{
dbi(drv, ...)
}
\arguments{
\item{drv}{passed to DBI::dbConnect}

\item{...}{optional args stored unevaluated and later evaluated just in time before passing to DBI::dbConnect}
}
\description{
Wrapper for DBI connection
}
