% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{default_cache}
\alias{default_cache}
\title{Get the default cache engine (currently \code{cache_memory()}).}
\usage{
default_cache()
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
Get the default cache engine (currently \code{cache_memory()}).
}
\seealso{
Other cache functions: \code{\link{cache_memory}}
}
\concept{cache functions}
