% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{default_eddy_env}
\alias{default_eddy_env}
\title{Get the default binding environment that keeps the eddies.}
\usage{
default_eddy_env()
}
\value{
An environment.
}
\description{
Get the default binding environment that keeps the eddies.
}
