% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{delete}
\alias{delete}
\title{Remove node}
\usage{
delete(x, ...)
}
\arguments{
\item{x}{node object or an id (character)}

\item{rflow}{rflow object}

\item{completely}{logical;}
}
\value{
delete methods can only return TRUE or raise error
}
\description{
Remove node
}
