% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{delete_eddy}
\alias{delete_eddy}
\title{Delete eddy and ALL its data from ALL cache layers}
\usage{
delete_eddy(cache_path = NULL, eddy_name = NULL,
  envir = get_default_env())
}
\description{
Delete eddy and ALL its data from ALL cache layers
}
