% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_primitives.R
\name{depends.node}
\alias{depends.node}
\title{List all dependencies recursively}
\usage{
\method{depends}{node}(x, inverse = FALSE, results = character())
}
\arguments{
\item{x}{a node object}

\item{inverse}{FALSE (default) => downtream, TRUE => upstream}

\item{results}{DO NOT USE}
}
\value{
vector of nodes' ids
}
\description{
List all dependencies recursively
}
