% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depends.R
\name{depends}
\alias{depends}
\title{List all dependencies recursively}
\usage{
depends(x, ...)
}
\arguments{
\item{x}{a node object}
}
\value{
vector of nodes' ids
}
\description{
List all dependencies recursively
}
