% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_deps.R
\name{detect_nodes}
\alias{detect_nodes}
\title{Detect dependencies}
\usage{
detect_nodes(x, envir = parent.frame(), space = "", st = 0,
  stmax = 20)
}
\arguments{
\item{space}{}
}
\description{
Detect dependencies
}
