% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_primitives.R
\name{eval_node}
\alias{eval_node}
\title{Evaluates a node}
\usage{
eval_node(x, ...)
}
\arguments{
\item{...}{args passed to node's eval() method}

\item{id}{node's id}
}
\description{
Evaluates a node assuming all its requirements are met
}
\details{
As oposed to \code{make()}, this function is not indended for frequent use by a user.
}
