% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{expand.vec}
\alias{expand.vec}
\title{Enforce recycling to a target length}
\usage{
expand.vec(vec, target.length)
}
\arguments{
\item{target.length}{}
}
\value{
vector \code{[target.length]}
}
\description{
Enforce recycling to a target length
}

