% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{expected.lines.RIARRAY}
\alias{expected.lines.RIARRAY}
\title{Number of lines in a MODFLOW text array}
\usage{
expected.lines.RIARRAY(header.txt, ncol, nrow, nlay = 1L)
}
\arguments{
\item{nlay}{}
}
\value{

}
\description{
Uses the header file to determine how many lines to read for a text
 array, excluding the header itself.
}
