% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{expression_r}
\alias{expression_r}
\title{Expression with source code references stripped off}
\usage{
expression_r(x)
}
\arguments{
\item{...}{expression (see \code{\link[base]{expression}})}
}
\value{
\code{expression_r} returns expression object similar to the one returned from \code{\link[base]{expression}} except references to source code
}
\description{
Expression with source code references stripped off
}
\examples{
identical(expression(1+1), expression_r(1+1))
identical(expression({1+1}), expression_r({1+1}))
}
\seealso{
\code{\link[base]{expression}}
}
