% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{flip}
\alias{flip}
\title{Flip an array}
\usage{
flip(arr, by)
}
\arguments{
\item{arr}{array}

\item{by}{integer \code{[1]}}
}
\value{
array
}
\description{
reflects an array in a given dimension
}
\examples{
mtx <- matrix(1:4, 2, 2)
mtx
Rflow:::flip(mtx, 1L)
Rflow:::flip(mtx, 2L)

}

