% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{flow}
\alias{flow}
\title{Implicit cache of a function and processing of the given call.}
\usage{
flow(fn_call, flow_options = get_flow_options())
}
\arguments{
\item{fn_call}{Function call to be processed.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The cached version of the function.
}
\description{
Implicit cache of a function and processing of the given call.
}
