% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{flow_call}
\alias{flow_call}
\title{Implicit cache of a function and of the given call.}
\usage{
flow_call(fn_call, fn_id = NULL, flow_options = get_flow_options())
}
\arguments{
\item{fn_call}{Function call to be processed.}

\item{fn_id}{Character or Integer. Optional id to uniquely identify 
the function. By default, rflow functions reuse the \code{cache} if the 
same function is given. The \code{fn_id} allows the user to suppress 
console messages and  to explicitly indicate whether to reuse the old
\code{cache} or create a new one.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The flow object.
}
\description{
Implicit cache of a function and of the given call.
}
\examples{
fn <- function(x, y) { x + y + 3 }
call_flow <- flow_call(fn(x = 1, y = 2))
collected_result <- call_flow \%>\% collect()

}
