% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-dfr.R
\name{flow_dfr}
\alias{flow_dfr}
\title{Row-wise caching of operations on data frame.}
\usage{
flow_dfr(df, ..., fn = NULL, fn_id = NULL,
  flow_options = get_flow_options())
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}. Rownames are not supported.}

\item{...}{Named arguments to pass to \code{fn}.}

\item{fn}{The function to apply to the data frame. It must accept a data
frame as the first argument and a numeric index as the second argument.}

\item{fn_id}{Optional id to uniquely identify the function. By default,
rflow functions reuse the cache if the same function is given. The id 
allows the user to suppress console messages and to explicitly
indicate whether to reuse the old cache or create a new one.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The flow object.
}
\description{
Row-wise caching of operations on data frame.
}
\details{
Function \code{fn} must return only the rows changed.
  The function \code{fn} may return fewer or more columns or modify 
  existing columns as long it always returns the same columns (data types 
  and names). This is consistent with the concept of "pure functions".
  The data frame \code{df} passed to \code{fn} will have one additional
  column \code{.row_hash} which must be return as is in order to identify
  changed rows.
}
