% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{flow_fn}
\alias{flow_fn}
\title{Explicit cache of a function.}
\usage{
flow_fn(fn, flow_options = get_flow_options())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The cached version of the function.
}
\description{
Explicit cache of a function.
}
