% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink-ns.R
\name{flow_ns_sink}
\alias{flow_ns_sink}
\title{Write a value to a namespace only if the value has changed.}
\usage{
flow_ns_sink(x, var_name, ns, flow_options = get_flow_options())
}
\arguments{
\item{x}{Value to assign.}

\item{var_name}{The name (as string) of the variable.}

\item{ns}{The name space, either an \code{environment} or a 
\code{Shiny::reactiveValues} object.}

\item{flow_options}{List of options created using \code{get_flow_options}.
All options except \code{excluded_arg} and \code{eddy} are ignored.}
}
\value{
The flow object
}
\description{
Write a value to a namespace only if the value has changed.
}
\examples{
fn <- function(x, y) { x + y }
flowed_fn <- flow_fn(1, 2, fn = fn)
sunk_flow <- flow_ns_sink(flowed_fn, "test_sink_flow", new.env()) 

}
