% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{flow_options}
\alias{flow_options}
\alias{default_flow_options}
\alias{set_flow_options}
\alias{get_flow_options}
\title{flow options used to initialize or update an eddy}
\usage{
default_flow_options(excluded_arg = character(),
  split_bare_list = TRUE, split_dataframe = FALSE, split_fn = NULL)

set_flow_options(excluded_arg = NULL, split_bare_list = NULL,
  split_dataframe = NULL, split_fn = NULL, eddy = get_current_eddy())

get_flow_options(excluded_arg = NULL, eval_arg_fn = NULL,
  split_bare_list = NULL, split_dataframe = NULL, split_fn = NULL,
  eddy = get_current_eddy())
}
\arguments{
\item{excluded_arg}{A vector of argument names to be excluded when computing
the input hash. Best used to exclude certain arguments that depend on 
the running state, e.g. a Shiny session, a parallel cluster, etc. 
Excluded arguments must have a default value to permit lazy computations.
The default is not to exclude any arguments from the input hash.}

\item{split_bare_list}{If the function output is a bare list 
(\code{\link[rlang:bare-type-predicates]{rlang::is_bare_list}}), determines
whether to calculate the hash of each list element and create
corresponding flow elements.}

\item{split_dataframe}{If the function output is a data.frame or tibble,
determines whether to calculate the hash of each column and create
corresponding flow elements.}

\item{split_fn}{Custom function to generate a list of elements from the
output of the flow-ed function. Useful only if the output is not a list
but a flow elements are still desired. Consider returning a list
as output before using this option.  If an \code{split_fn}
is provided, \code{split_bare_list} and \code{split_dataframe} will be
ignored.}

\item{eddy}{Eddy to apply / retrieve options to / from.}

\item{eval_arg_fn}{Custom function to parse the input arguments and create
a list of evaluated arguments to be hashed. This function should have the
exact same arguments as the original function. Try to use 
\code{excluded_arg} or flow source before creating a custom function. 
Because each custom function is flow specific, it is not possible to set 
this option at the eddy level using \code{set_flow_options}.}
}
\value{
For \code{default_flow_options}, a list of options.

For \code{set_flow_options}, \code{NULL}.

For \code{get_flow_options}, a list of options 
including the eddy.
}
\description{
flow options used to initialize or update an eddy

default flow options used to initialize or update an eddy.
}
\details{
If used in \code{set_flow_options}, these options will be stored in 
  \code{eddy} and retrieved by each flow subsequently executed. E.g. if not
  careful, it is possible to force all following flows to use a custom 
  \code{split_fn} function, which is not recommended.

Sets a list of default options.

\code{set_flow_options} does not overwrite the current options when the 
  argument is \code{NULL}.
}
