% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gcs}
\alias{gcs}
\alias{gccs}
\alias{grcs}
\alias{mids}
\title{Grid line co-ordinates}
\usage{
gccs(data, absolute = FALSE, x0)

grcs(data, absolute = FALSE, y0)

mids(v)
}
\arguments{
\item{data}{DIS.MFpackage or NetCDF object;
MODFLOW data relating to the model}

\item{absolute}{logical \code{[1]};
\code{TRUE} if the absolute, or global co-ordinates should be returned.
 The default is \code{FALSE}, so that the first co-ordinate is always 0}

\item{x0, y0}{numeric \code{[1]};
If \code{data} is a DIS package and \code{absolute = TRUE}, then the
 value of \eqn{x} (\code{gccs}) or \eqn{y} (\code{grcs}) at the model
 origin (bottom left corner) must be given.  Not needed with NetCDFs
 because this information is stored in the NetCDF output generated from
 \code{Rflow}.}

\item{v}{numeric \code{[]};
a vector of grid divider co-ordinates}
}
\value{
\code{gccs}: numeric \code{[NCOL + 1]}\cr
\code{grcs}: numeric \code{[NROW + 1]}\cr
\code{mids}: numeric \code{[length(v) - 1]}
}
\description{
Find the co-ordinates of column (\code{gccs}) or row (\code{grcs})
 dividers from the DIS package or the NetCDF-format output data.
}
\details{
\code{gccs} and \code{grcs} return the x co-ordinates of the column
 dividers and the y co-ordinates of the row dividers respectively, in
 each case in ascending order, which will be the reverse of the row
 order in the latter case.  Wrap with \code{mids} to find the midpoints,
 that is the node centres.
}
\examples{
# using DIS.MFpackage
dis <- read.DIS(system.file("rflow_mf_demo.dis",
                            package = "Rflow"))
gccs(dis)
grcs(dis)

# arbitrary origin
gccs(dis, TRUE, 1000)

# using NetCDF
mfdata <- RNetCDF::open.nc({
  system.file("rflow_mf_demo.nc", package = "Rflow")
})

gccs(mfdata)
grcs(mfdata)

# origin is in fact 0,0 for the demo, so these will show no difference
gccs(mfdata, TRUE)
grcs(mfdata, TRUE)

# - if t0 is given with a NetCDF, a warning is given to avoid
#    ambiguity
\dontrun{gccs(mfdata, TRUE, 1000)}

# node centre x and y co-ordinates:
# - x
mids(gccs(mfdata, TRUE))
# - y
mids(grcs(mfdata, TRUE))

}
