% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{generate_code}
\alias{generate_code}
\title{Generate R code from nodes}
\usage{
generate_code(con)
}
\arguments{
\item{con}{Connection object for the communication server}
}
\value{
Character scalar holding the generated code
}
\description{
Takes a connection to the node app and inserts R code corresponding to nodes
}

