% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{generate_nodes}
\alias{generate_nodes}
\title{Generate a JSON Representation of User-facing Functions}
\usage{
generate_nodes(pkg)
}
\arguments{
\item{pkg}{Character scalar naming a package}
}
\description{
Generates a JSON with function names, arguments, default values, and
  manuals for all user-facing functions in a package
}

