% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtop.R
\name{get.wtop.nc}
\alias{get.wtop.nc}
\title{Phreatic groundwater top}
\usage{
get.wtop.nc(mfdata, fnm, nts.dtit = "NTS")
}
\arguments{
\item{mfdata}{NetCDF object (see \code{\link{GW.nc}});
MODFLOW results}

\item{fnm}{character string;
file name to save to}

\item{nts.dtit}{\code{"NTS"} (almost always) or \code{"sNTS"};
name of the dimension giving the number of time steps (\code{"sNTS"}) is
 used when the NetCDF has been split (see \code{\link{GW.nc}})}
}
\value{
NetCDF object with "wtop" variable representing the top of saturated
 groundwater in each cell, either the water table or the cell top,
 whichever is lower
}
\description{
Phreatic groundwater top
}
\examples{
mfdata <- RNetCDF::open.nc(system.file("rflow_mf_demo.nc", package = "Rflow"))

wtop <- get.wtop.nc(mfdata, "RFLOW_EXAMPLE_wtop.nc")
print.nc(wtop)
}
