% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nodes.R
\name{get_args}
\alias{get_args}
\title{Get Arguments of User-facing Functions}
\usage{
get_args(fun_name, pkg)
}
\arguments{
\item{fun_name}{Character scalar naming an object in a package's namespace}

\item{pkg}{Character scalar naming a package}
}
\value{
list of data_frames holding a column of argument names and a column of 
  default values
}
\description{
For each user-facing function in a package, obtain argument names
  and default values
}

