% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{get_current_eddy}
\alias{get_current_eddy}
\title{Get the current eddy for a given (or default) environment.}
\usage{
get_current_eddy(eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_env}{An environment where to put (bind) the eddy.}
}
\value{
An eddy object to be used for storing rflows.
}
\description{
If the current eddy was not previously set with \code{set_current_eddy},
  it creates a new eddy that uses \code{default_cache()}.
}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_eddy}}, \code{\link{new_eddy}},
  \code{\link{set_current_eddy}}
}
\concept{eddy functions}
