% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{get_dag}
\alias{get_dag}
\title{Look up a previously created DAG by id from the in-session registry}
\usage{
get_dag(dag_id)
}
\arguments{
\item{dag_id}{character}
}
\value{
a [DAG] object
}
\description{
Look up a previously created DAG by id from the in-session registry
}
