% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{get_default_eddy}
\alias{get_default_eddy}
\title{Get the eddy associated with the default environment.}
\usage{
get_default_eddy(cache_path = NULL, envir = get_default_env())
}
\value{
An R6Eddy object.
}
\description{
Get the eddy associated with the default environment.
}
\seealso{
\code{\link{get_default_env}}
}
