% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{get_default_env}
\alias{get_default_env}
\title{Get the default environment that keeps the eddies.}
\usage{
get_default_env()
}
\value{
An environment object.
}
\description{
Get the default environment that keeps the eddies.
}
\seealso{
\code{\link{get_default_eddy}}
}
