% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nodes.R
\name{get_docs}
\alias{get_docs}
\title{Get Documentation for a User-facing Function}
\usage{
get_docs(fun_name, pkg)
}
\arguments{
\item{fun_name}{Character scalar naming an object in a package's namespace}

\item{pkg}{Character scalar naming a package}
}
\value{
List of html formatted manuals of user-facing functions
}
\description{
For a user-facing function in some package, retrieve its manual
  and convert it into html format
}

