% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{get_eddy}
\alias{get_eddy}
\title{Retrieves an eddy.}
\usage{
get_eddy(eddy_name, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to allow retrieving later.}

\item{eddy_env}{An environment where to put (bind) the eddy.}
}
\value{
An eddy object to be used for storing rflows.
}
\description{
Retrieves an eddy.
}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_current_eddy}}, \code{\link{new_eddy}},
  \code{\link{set_current_eddy}}
}
\concept{eddy functions}
