% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nodes.R
\name{get_funs}
\alias{get_funs}
\title{Get Names of User-facing Functions}
\usage{
get_funs(pkg)
}
\arguments{
\item{pkg}{Character scalar naming a package}
}
\value{
Character vector holding function names
}
\description{
Using user-facing tests, obtain function names from a package namespace
}

