% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{get_id}
\alias{get_id}
\title{extract id from a object's definition (in a list)}
\usage{
get_id(obj)
}
\arguments{
\item{obj}{object definition}
}
\description{
extract id from a object's definition (in a list)
}
