% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_primitives.R
\name{get_nodes_ids}
\alias{get_nodes_ids}
\title{Make a vector of all nodes' ids}
\usage{
get_nodes_ids(x)
}
\arguments{
\item{x}{an rflow object}
}
\value{
Character vector of all nodes' ids
}
\description{
Make a vector of all nodes' ids
}
