% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{get_value}
\alias{get_value}
\title{Obtain value represented by a node}
\usage{
get_value(id, rflow)
}
\arguments{
\item{id}{node's id}

\item{rflow}{rflow object}
}
\value{
Value/object represented by given node.
}
\description{
Obtain value represented by a node
}
