% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_ncdf.R
\name{GW.nc}
\alias{GW.nc}
\title{MODFLOW results to NetCDF data set}
\usage{
GW.nc(dir, mfrt, ncrt = mfrt, DIS = paste0(mfrt, ".dis"),
  HDS = paste0(mfrt, ".hds"), CBB = paste0(mfrt, ".cbb"),
  CBW = paste0(mfrt, ".cbw"), CRC = paste0(mfrt, ".crc"),
  CRV = paste0(mfrt, ".crv"), CBD = paste0(mfrt, ".cbd"),
  CBG = paste0(mfrt, ".cbg"), CS1 = paste0(mfrt, ".cs1"),
  BAS = paste0(mfrt, ".bas"), LPF = paste0(mfrt, ".lpf"),
  BCF = paste0(mfrt, ".bcf"), HNOFLO, HDRY,
  title = "untitled MODFLOW data set", author = "", datum = "",
  length.unit = "metre", time.unit = "day", xyt0 = c(0, 0, 0),
  start_date = "", updating = TRUE, .FillValue = 1e+30,
  split.tss = NULL, ch = TRUE)
}
\arguments{
\item{dir}{character string;
directory in which reside the MODFLOW output and input files}

\item{mfrt}{character string;
non-extended filename to use as root for finding default MODFLOW input and
 output file names}

\item{ncrt}{character string;
filename (minus extension) for resulting NetCDF data set (default is \code{mfrt})}

\item{DIS, BAS, LPF, BCF}{character string;
File names for DIS, BAS, LPF and BCF package files.  The latter three
 are only used for finding \code{HNOFLO} and \code{HDRY} if these are
 not given.}

\item{HDS}{character string;
File name for head array.}

\item{CBB, CBW, CRC, CRV, CBD, CBG, CS1}{character string;
File names for budget files, cell-by-cell flows and those relating to
 specific boundary conditions.  If a file doesn't exist it is simply
 ignored.}

\item{HNOFLO}{numeric \code{[1]};
value to interpret as inactive cells in head array, automatically gleaned
 from BAS package file if not given}

\item{HDRY}{numeric \code{[1]};
value to interpret as dry cell in head array, automatically gleaned from LPF
 or BCF package file if not given}

\item{title}{character string;
title of the model, optional but recommended}

\item{author}{character string;
who constructed the MODFLOW model?  (Not who performed the conversion to
 NetCDF.)}

\item{datum}{character string;
description of the reference datum for height/ head values, for example
 "Ordnance Datum"}

\item{length.unit, time.unit}{character string;
implied length and time units of the model, recommended to use full name
 (e.g. \code{"metre"} not \code{"m"})}

\item{xyt0}{numeric \code{[3]};
global x and y values at the bottom left corner of the model and the absolute
 time value that the start of the model refers to}

\item{start_date}{character string;
purely explanatory, enables the author to describe what start_time refers to}

\item{updating}{logical \code{[1]};
Get meta data (global attributes) from existing NetCDF data set?  Will
 only attempt if the NetCDF data set already exists, so the default is
 TRUE.  This can save rewriting meta data such as title and origin.}

\item{split.tss}{integer \code{[]};
time steps at which to split into multiple NetCDFs; the end time step of the
 split group is to be given; an integer vector if this feature is to be used}

\item{ch}{logical \code{[1]};
whether to save the constant head flux (in a simulation with no CHs, there
 will still be arrays for constant head in the budget files, which waste space
 as all the elements are 0; it is not possible to know \emph{a priori} whether
 the CH array is needed, so the user should say if s/he knows)}
}
\value{
\code{NULL}

Prints a summary of the final NetCDF data set using print.nc.
}
\description{
MODFLOW results to NetCDF data set
}
\examples{
# find the directory
wkdir <- paste0(system.file("inst", package = "Rflow"))

# make fresh
GW.nc(wkdir, "rflow_mf_demo", "RFLOW_EXAMPLE", title = "example",
      author = "model builder's name", datum = "OD or something",
      xyt0 = c(0, 0, 0), start_date = "the start time descriptively")

# update results, but keep metadata, with updating = TRUE (default)
GW.nc(wkdir, "rflow_mf_demo", "RFLOW_EXAMPLE")

library(RNetCDF)
mfdata <- open.nc(paste0(wkdir, "/rflow_mf_demo.nc"))
print.nc(mfdata)

}
