% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{handler_file}
\alias{handler_file}
\title{Handler constructors}
\usage{
handler_file(path, open = TRUE, enable = TRUE)
}
\arguments{
\item{path}{to a log file}

\item{open}{logical; open the connection immediately}

\item{enable}{logical; enable the handler}
}
\description{
Handler constructors
}
