% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in.R6}
\alias{in.R6}
\title{\code{\%in\%} equivalent for R6 objects}
\usage{
in.R6(obj, objs)
}
\arguments{
\item{obj}{R6 object}

\item{objs}{vector of R6 objects}
}
\description{
Help function for matching R6 objects
}
