% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{interpret.RIARRAY}
\alias{interpret.RIARRAY}
\title{Convert lines of text to a numeric array}
\usage{
interpret.RIARRAY(txt.lines, header = T, ncol, nlay = 1L)
}
\arguments{
\item{nlay}{}
}
\value{

}
\description{
Convert lines of text to a numeric array
}
