% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{is.comment}
\alias{is.comment}
\title{Check if line of text is a comment}
\usage{
is.comment(line, comm.char = "#", allow.ws = FALSE)
}
\arguments{
\item{line}{character string}

\item{comm.char}{character string;
character (or string is permissible) to identify comment line}

\item{allow.ws}{logical \code{[1]};
if FALSE, the comment character must be the first character, including
 spaces, for the line to be identified as a comment}
}
\value{
logical \code{[1]}
}
\description{
Check if line of text is a comment
}
\examples{
Rflow:::is.comment(" #  ")
Rflow:::is.comment(" #  ", allow.ws = TRUE)
Rflow:::is.comment("! test", "!")
Rflow:::is.comment("code", "")

}
