% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_computed}
\alias{is_computed}
\title{Is the current state computed (stored in the cache)?}
\usage{
is_computed(rflow)
}
\arguments{
\item{rflow}{An rflow object, e.g. as returned by \code{\link{rf}}.}
}
\value{
A logical value, whether the value can be obtained without
  triggering the computation chain.
}
\description{
Is the current state computed (stored in the cache)?
}
