% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_current}
\alias{is_current}
\title{Does the flow have a "current" state?}
\usage{
is_current(flow)
}
\arguments{
\item{flow}{An flow object, e.g. as returned by \code{\link{flow_call}}.}
}
\value{
A logical value, whether the current state is valid.
}
\description{
If there is no current state, e.g. right after \code{\link{make_flow_fn}},
  the flow is "not flowing", it is preventing downstream flows
  from being computed.
}
