% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{is_dag.rflow}
\alias{is_dag.rflow}
\title{Checks a graph as represented by Rflow is a acyclic (DAG)}
\usage{
is_dag.rflow(x)
}
\arguments{
\item{x}{an Rflow object}
}
\value{
TRUE is the Rflow is a DAG
}
\description{
Checks a graph as represented by Rflow is a acyclic (DAG)
}
\examples{
\dontrun{
is_dag(MYRFLOW)
}
}
