% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_flow_fn}
\alias{is_flow_fn}
\title{Checks if the function is flow function (as returned by \code{make_flow_fn})}
\usage{
is_flow_fn(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
A logical value, whether \code{fn} is a flow function.
}
\description{
Checks if the function is flow function (as returned by \code{make_flow_fn})
}
\examples{
fn <- function(x, y) { x + y + 12 }
flowed_function <- flow_fn(2, 3, fn = fn)
is_flow_function <- is_flow_fn(fn = flowed_function)

}
