% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_not_flow_fn}
\alias{is_not_flow_fn}
\title{Check if the function is NOT a flow function
(as returned by \code{make_flow_fn}).}
\usage{
is_not_flow_fn(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
A logical value, whether \code{fn} is a not flow function.
}
\description{
Check if the function is NOT a flow function
(as returned by \code{make_flow_fn}).
}
