% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_valid}
\alias{is_valid}
\title{Is the current state valid (stored in the cache)?}
\usage{
is_valid(flow, state = "current")
}
\arguments{
\item{flow}{An flow object, e.g. as returned by \code{\link{flow_call}}.}

\item{state}{An flow state. It can be either a valid state 
\code{index} (integer) or a valid state: \code{"current"}, \code{"all"}, 
\code{in_hash} or \code{out_hash} (string).}
}
\value{
A logical value, whether the value can be obtained without
  triggering computation.
}
\description{
Is the current state valid (stored in the cache)?
}
