% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{job}
\alias{job}
\title{jobs
\itemize{
\item can be run = must evaluate() method
\item can have dependencies (nodes)
\item running them leads to consequences (nodes) that can be verified
}}
\usage{
job(x, ...)
}
\description{
jobs
\itemize{
\item can be run = must evaluate() method
\item can have dependencies (nodes)
\item running them leads to consequences (nodes) that can be verified
}
}
