% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FilterWith}
\alias{FilterWith}
\alias{lapplyWith}
\alias{sapplyWith}
\title{Filter with expressions}
\usage{
FilterWith(x, ...)

lapplyWith(x, ...)

sapplyWith(x, ...)
}
\arguments{
\item{x}{a vector of data (typically a list of similar data objects)}

\item{...}{passed to with()}

\item{expr}{expression passed to with()}
}
\value{

}
\description{
Filter with expressions
}
