% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_runs}
\alias{list_runs}
\title{Print a summary table of DAG runs}
\usage{
list_runs(con, dag_id = NULL)
}
\arguments{
\item{con}{connection from [rflow_db_connect()]}

\item{dag_id}{optional filter}
}
\description{
Print a summary table of DAG runs
}
