% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{load_nodes.rflow}
\alias{load_nodes.rflow}
\title{batch load object definitions}
\usage{
\method{load_nodes}{rflow}(x, conflict = "update", verbose = TRUE)
}
\arguments{
\item{conflict}{logical; How to resolve conflict when an object of the same id already exists in the rflow?}

\item{verbose}{logical; print verbose output?}

\item{rflow}{an rflow object}
}
\description{
batch load object definitions
}
