% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{load_obj_definitions}
\alias{load_obj_definitions}
\title{load DAG's objects' definition from TOML files}
\usage{
load_obj_definitions(path, modified_since = NULL, verbose = TRUE)
}
\arguments{
\item{path}{path to folder with node definitions}
}
\description{
load DAG's objects' definition from TOML files
}
