% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{load_objs_state}
\alias{load_objs_state}
\title{batch loading objects' definition/state from RDS files}
\usage{
load_objs_state(path, recursive = FALSE, ignore.case = TRUE, ...)
}
\description{
batch loading objects' definition/state from RDS files
}
