% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{load_state_of_node}
\alias{load_state_of_node}
\title{loads object definition/state from a RDS file}
\usage{
load_state_of_node(path, ...)
}
\arguments{
\item{path}{path to RDS file with saved state}
}
\description{
loads object definition/state from a RDS file
}
