% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_record}
\alias{log_record}
\alias{log_record.logger}
\alias{log_record.rflow}
\alias{log_record.list}
\title{Log a record/message}
\usage{
log_record(x, ...)

\method{log_record}{logger}(logger, ...)

\method{log_record}{rflow}(rflow, ...)

\method{log_record}{list}(x, ...)
}
\arguments{
\item{x}{an object of class that supprt logging}

\item{...}{args passed to a logger}
}
\description{
Log a record/message
}
