% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logger}
\alias{logger}
\title{logger constructor}
\usage{
logger(name = "ROOT", enabled = TRUE, handlers = NULL)
}
\arguments{
\item{name}{name of the logger}

\item{enabled}{logical; enable the logger on initialization?}

\item{handlers}{a list of handlers can be supplied}
}
\description{
logger constructor
}
