% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{make}
\alias{make}
\title{Make/build target or multiple targets}
\usage{
make(x, ...)
}
\arguments{
\item{x}{node's id or rflow object}

\item{force}{logical; force eval()}

\item{verbose}{logical; switch on/off verbose output}
}
\description{
Make/build target or multiple targets
}
