% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{make_file_source}
\alias{make_file_source}
\title{Cache a function - TODO.}
\usage{
make_file_source(fn, file_arg = 1, split_output_fn = NULL,
  eddy = get_default_eddy())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{file_arg}{TODO.}

\item{split_output_fn}{Custom function to split the output into a list,
in order to hash its elements separately.}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the function.
}
\description{
Cache a function - TODO.
}
