% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{make_flow_fn}
\alias{make_flow_fn}
\title{Explicit cache of a function.}
\usage{
make_flow_fn(fn, fn_id = NULL, flow_options = get_flow_options())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{fn_id}{Optional id to uniquely identify the function. By default,
rflow functions reuse the \code{cache} if the same function is given. The id 
allows the user to suppress console messages and to explicitly
indicate whether to reuse the old \code{cache} or create a new one.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The cached version of the function.
}
\description{
Explicit cache of a function.
}
\details{
In order to use the functionality of an \code{R6Flow} object, the output of 
\code{make_flow_fn} has to be collected first.
}
\examples{

input_function <- function(x, y) { x + y }
make_flow_function <- make_flow_fn(input_function, fn_id = 1) 
rflow_function <- make_flow_function(2, 3)
flow_result <- rflow_function \%>\% collect()

# usage with rflow pipes
input_function2 <- function(x, y) { x * y }
make_flow_function2 <- make_flow_fn(input_function2, fn_id = 2) 
collected_pipe_flow <- make_flow_function(1, 2) \%>\%
                     make_flow_function2(2) \%>\%
                     collect()

}
