% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{make_ns_sink}
\alias{make_ns_sink}
\title{Cache a function - TODO.}
\usage{
make_ns_sink(var, ns, hash_input_fn = NULL, eddy = get_default_eddy())
}
\arguments{
\item{var}{The name (as string) of the variable.}

\item{ns}{The name space, either an \code{environment} or a 
\code{Shiny reactiveValue} object.}

\item{hash_input_fn}{Custom function to process only a part of the input
(e.g. skip one fn inputs - not recommended).}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the function.
}
\description{
Cache a function - TODO.
}
