% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{make_rf}
\alias{make_rf}
\title{Explicit cache of a function.}
\usage{
make_rf(fn, rflow_options = get_rflow_options())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{rflow_options}{List of options created using \code{get_rflow_options}.}
}
\value{
The cached version of the function.
}
\description{
Explicit cache of a function.
}
