% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{make_rflow}
\alias{make_rflow}
\title{Cache a function.}
\usage{
make_rflow(fn, hash_input_fn = NULL, split_output_fn = NULL,
  eddy = get_default_eddy())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{hash_input_fn}{Custom function to process only a part of the input
(e.g. skip one fn inputs - not recommended).}

\item{split_output_fn}{Custom function to split the output into a list,
in order to hash its elements separately.}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the function.
}
\description{
Cache a function.
}
