% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{make_sink}
\alias{make_sink}
\title{Cache a function - TODO.}
\usage{
make_sink(fn, hash_input_fn = NULL, eddy = get_current_eddy())
}
\arguments{
\item{fn}{Function to be cached, ideally a pure function.}

\item{hash_input_fn}{Custom function to process only a part of the input
(e.g. skip one fn inputs - not recommended).}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the function.
}
\description{
Cache a function - TODO.
}
