% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplot.R
\name{MFimage}
\alias{MFimage}
\title{Plot MODFLOW variable matrix}
\usage{
MFimage(mtx, gccs, grcs, zlim = "auto", col = colorRampPalette(c("blue",
  "white", "red"))(101L), show.range = identical(zlim, "angle"), ...)
}
\arguments{
\item{mtx}{matrix;
variable to plot}

\item{gccs}{numeric \code{[NCOL + 1]};
column divider \eqn{x} co-ordinates, in ascending order}

\item{grcs}{numeric \code{[NROW + 1]};
row divider \eqn{y} co-ordinates, in ascending order}

\item{zlim}{numeric \code{[2]} or character string;
min. and max. values to plot; alternatively use \code{breaks} argument
 as in \code{graphics::image}; additionally, three character string
 options are given: \code{"auto"} for automatic range calculation
 (pretty), \code{"sym"} for automatic range that is symmetric about \eqn{0}
 and "angle" for a range between \eqn{-pi} and \eqn{pi}}

\item{col}{character \code{[]};
colours}

\item{show.range}{logical \code{[1]};
print computed \code{zlim} to screen? (in case that
 \code{zlim = "angle"}, shows which directions the colours represent)}

\item{...}{additional plotting arguments for \code{image}}
}
\description{
Plot MODFLOW variable matrix
}
\examples{
# plot model domain
fnms <- system.file(c("rflow_mf_demo.bas",
                      "rflow_mf_demo.dis"),
                    package = "Rflow")

dis <- read.DIS(fnms[2L])
bas <- read.BAS(fnms[1L], dis)

# grid divider co-ordinates (the example has a regular grid)
gccs <- with(dis, seq(0, by = DELR, length.out = extent["NCOL"] + 1L))
grcs <- with(dis, seq(0, by = DELC, length.out = extent["NROW"] + 1L))

MFimage(bas$IBOUND[,, 1L], gccs, grcs, c(-1, 1),
        c("blue", "grey", "transparent"))

}
