% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplot.R
\name{MFncimage}
\alias{MFncimage}
\title{Plot MODFLOW variable matrix from NetCDF}
\usage{
MFncimage(nc, variable, start = c(NA, NA, 1L, 1L), count = c(NA, NA, 1L,
  1L), zlim = "auto", col = colorRampPalette(c("blue", "white",
  "red"))(101L), show.range = identical(zlim, "angle"), ...)
}
\arguments{
\item{nc}{NetCDF object;
an open connection to a MODFLOW/ MT3DMS data set}

\item{variable}{character string;
name of the variable to be plotted}

\item{start}{integer \code{[4]};
index in each dimension at which to start reading (see
 \code{\link{var.get.nc, pkg = RNetCDF}}); \code{NA} implies start}

\item{count}{integer \code{[4]};
numbers of indices to read in each dimension (see
 \code{\link{var.get.nc, pkg = RNetCDF}}); \code{NA} implies to end}

\item{zlim}{numeric \code{[2]} or character string;
min. and max. values to plot; alternatively use \code{breaks} argument
 as in \code{graphics::image}; additionally, three character string
 options are given: \code{"auto"} for automatic range calculation
 (pretty), \code{"sym"} for automatic range that is symmetric about \eqn{0}
 and "angle" for a range between \eqn{-pi} and \eqn{pi}}

\item{col}{character \code{[]};
colours}

\item{show.range}{logical \code{[1]};
print computed \code{zlim} to screen? (in case that
 \code{zlim = "angle"}, shows which directions the colours represent)}

\item{...}{}

\item{...}{Arguments passed on to \code{MFimage}
\describe{
  \item{mtx}{matrix;
variable to plot}
  \item{gccs}{numeric \code{[NCOL + 1]};
column divider \eqn{x} co-ordinates, in ascending order}
  \item{grcs}{numeric \code{[NROW + 1]};
row divider \eqn{y} co-ordinates, in ascending order}
  \item{zlim}{numeric \code{[2]} or character string;
min. and max. values to plot; alternatively use \code{breaks} argument
 as in \code{graphics::image}; additionally, three character string
 options are given: \code{"auto"} for automatic range calculation
 (pretty), \code{"sym"} for automatic range that is symmetric about \eqn{0}
 and "angle" for a range between \eqn{-pi} and \eqn{pi}}
  \item{col}{character \code{[]};
colours}
  \item{show.range}{logical \code{[1]};
print computed \code{zlim} to screen? (in case that
 \code{zlim = "angle"}, shows which directions the colours represent)}
}}
}
\description{
Plot MODFLOW variable matrix from NetCDF
}
\examples{

library("RNetCDF")
fnm <- system.file("rflow_mf_demo.nc", package = "Rflow")
mfdata <- open.nc(fnm)

# plot the flux from river (blue is negative)
MFncimage(mfdata, "RiverLeakage",
          c(NA, NA, 1L, 10L),
          c(NA, NA, 1L, 1L),
          "sym", show.range = TRUE)

}
