% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{mftime}
\alias{mfsptime}
\alias{mftime}
\alias{mftstime}
\title{Time Step and Stress Period timings}
\usage{
mftstime(data, absolute = FALSE, t0)

mfsptime(data, absolute = FALSE, t0)
}
\arguments{
\item{data}{DIS.MFpackage or NetCDF object;
mfsptime does not yet support NetCDF}

\item{absolute}{logical \code{[1]};
\code{TRUE} if the absolute time values should be returned.  The default
 is \code{FALSE}.}

\item{t0}{numeric \code{[1]};
If \code{data} is a DIS package and \code{absolute = TRUE}, then the
 value of \eqn{t} at the start of the model must be given.  Not needed
 with NetCDFs because this information is stored in the NetCDF output
 generated from \code{Rflow}.}
}
\value{
\code{mftstime}:\cr
numeric \code{[NTS + 1]}, where \code{NTS} is the number of time steps;
the time values at the time step transitions, including the start and end
 times of the model

\code{mfsptime}:\cr
numeric \code{[NPER + 1]}, where \code{NPER} is the number of stress
periods; the time values at the stress period transitions, including the
 start and end times of the model
}
\description{
Get the time values of the time step or stress period divides from a DIS
 package or a MODFLOW NetCDF data set.  Stress periods define periods of
 different boundary condition stresses.  Stress periods are divided into
 multiple time steps in order to describe the transient response
 accurately.
}
\examples{
# using DIS.MFpackage
dis <- read.DIS(system.file("rflow_mf_demo.dis",
                            package = "Rflow"))

mftstime(dis)
mfsptime(dis)

# - with a made-up start time for demonstration purposes
mftstime(dis, TRUE, 1000)
mfsptime(dis, TRUE, 1000)

# using a NetCDF
mfdata <- RNetCDF::open.nc({
  system.file("rflow_mf_demo.nc", package = "Rflow")
})

mftstime(mfdata)

# - the start time is in fact 0 for the demo, so this will show no
#    difference
mftstime(mfdata, TRUE)

# - if t0 is given with a NetCDF, a warning is given to avoid
#    ambiguity
\dontrun{mftstime(mfdata, TRUE, 1000)}

}

