% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_ncdf.R
\name{MT3DMS.nc}
\alias{MT3DMS.nc}
\title{MT3DMS results to NetCDF}
\usage{
MT3DMS.nc(dir, mtrt, ntts = Inf, gw.nc, file = paste0(mtrt, ".nc"),
  Nspecies = 10L, species.names = as.character(1:Nspecies),
  files = cbind(mob = paste0(mtrt, 1:Nspecies, ".ucn"), immob = paste0(mtrt,
  "_sorbed", 1:Nspecies, ".ucn")), CNOFLO = -1,
  title = "untitled MT3DMS data set", author = "", conc.unit = "kg/m^3",
  length.unit = "metre", time.unit = "day", updating = TRUE)
}
\arguments{
\item{dir}{character string;
directory in which input and output files are stored}

\item{mtrt}{character string;
default file name, minus extensions, for the data sets}

\item{ntts}{integer \code{[]};}

\item{updating}{}
}
\value{
\code{NULL}
}
\description{
MT3DMS results to NetCDF
}
\examples{

}
