% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{nicearname}
\alias{nicearname}
\title{Format array headers from MODFLOW}
\usage{
nicearname(arty)
}
\arguments{
\item{arty}{character string}
}
\value{
character string
}
\description{
Format array headers from MODFLOW
}
\examples{
Rflow:::nicearname(" FLOW RIGHT FACE") # "FlowRightFace"

}

