% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{class}
\name{node}
\alias{node}
\title{node class}
\format{\code{\link{R6Class}} object.}
\usage{
node$new(id = "node")
}
\value{
Object of \code{\link{R6Class}} with basic functionality for rflows.
}
\description{
Node is a basic building block of an rflow DAG.
}
\details{
Nodes may represent different types of data targets (R objects, files, database tables), or jobs.
Interconnected nodes together function as basic building elements of directed acyclic graph (DAG).

Nodes are implemented using R6 classes. There are various flavours of the basic node class (r_node, db_node, file_node) that have their specific requirements and features

Constructing nodes
\itemize{
\item from stored state
\item from lists (comming from TOML, YAML...)
}
}
\section{Fields}{

\describe{
\item{\code{id}}{character; unique id}

\item{\code{env}}{character; name of a environemnt (container or group of objects)}

\item{\code{name}}{character; object's name}

\item{\code{desc}}{character; description}

\item{\code{depends}}{character; vector of uppstream nodes (other nodes this node is depending on)}
}}

\section{Methods}{

\describe{
\item{\code{$make()} or \code{make(node)}}{This method is used to build/make targets. It recursively solves dependencies by building all targets for nodes declared or detected as a dependency.}
\item{\code{$eval()}}{Compared to \code{make()}, \code{eval()} only runs the code/job associated with the node alone. It does not try to ensure all the dependencies are ready. \code{eval()} is usually used internally by \code{make()}.}
\item{\code{$value}}{Returns target's value. Targets such as database tables or csv files are automatically convert into R values.}
\item{\code{$exists()}}{Checks whether the target exists or not.}
}
}

\keyword{node}
