% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{node_definitions}
\alias{node_definitions}
\title{Construct a list of nodes' definitions}
\usage{
node_definitions(...)
}
\arguments{
\item{...}{arguments packed into a list of definitions}
}
\description{
Construct a list of nodes' definitions
}
