% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_primitives.R
\name{nodes}
\alias{nodes}
\title{Returns a list with all node objects}
\usage{
nodes(x)
}
\arguments{
\item{x}{an rflow object}
}
\value{
List of all node object from given rflow
}
\description{
Returns a list with all node objects
}
