% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{nodes_from_r_script}
\alias{nodes_from_r_script}
\title{Load nodes' definitions from an R script.}
\usage{
nodes_from_r_script(file, encoding = "UTF-8", ...)
}
\arguments{
\item{file}{path to R script file}

\item{encoding}{passed to parse()}

\item{...}{args passed to source()}
}
\description{
Load nodes' definitions from an R script.
}
