% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{numeric.extract}
\alias{numeric.extract}
\title{Extract numerical values from a string}
\usage{
numeric.extract(x, sep = " ")
}
\arguments{
\item{x}{character string;}

\item{sep}{character;
0 or 1 characters}
}
\value{
numeric \code{[]}
}
\description{
Extract numerical values from a string
}
\examples{
# sep = " " insists on exactly one space between numbers
Rflow:::numeric.extract("1 2  3", " ")

# sep = "" allows any amount of white space between numbers
Rflow:::numeric.extract("1 2  3", "")

}

