% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rflow}
\alias{plot.rflow}
\title{visualize Rflow DAGs using visNetwork}
\usage{
\method{plot}{rflow}(rflow, direction = "LR", ...)
}
\arguments{
\item{rflow}{an rflow objects}

\item{direction}{see visNetwork docs on hierarchical graphs}

\item{...}{args passed to visNetwork}
}
\description{
visualize Rflow DAGs using visNetwork
}
