% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{Prepare Text Data}
\alias{Prepare Text Data}
\alias{prepareTextData}
\title{Prepare Text Data}
\usage{
prepareTextData(fileName, dataType = "train", targetColInd = 0,
  predictorColInd = 2, delimiter = ",")
}
\arguments{
\item{fileName}{Path to the data}

\item{dataType}{Whether this is 'train' or 'test' data}

\item{targetColInd}{The column index of the target variable (labels)}

\item{predictorColInd}{The column index of the predictor (text)}

\item{delimiter}{The delimiter used to parse the data}
}
\description{
This function loads/prepares data for text classification
}

