% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{Preparing predictors}
\alias{Preparing predictors}
\alias{preparePredictors}
\title{Preparing predictors}
\usage{
preparePredictors(predictors)
}
\arguments{
\item{predictors}{A data.frame/data.table following R's convention}
}

