% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{Preparing target variable}
\alias{Preparing target variable}
\alias{prepareTargetVar}
\title{Preparing target variable}
\usage{
prepareTargetVar(target)
}
\arguments{
\item{target}{A data.frame/data.table following R's convention}
}

