% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{print_with_prefix}
\alias{print_with_prefix}
\title{Pretty printing of R expressions}
\usage{
print_with_prefix(x, verbose_prefix = "")
}
\arguments{
\item{verbose_prefix}{}
}
\description{
Pretty printing of R expressions
}
