% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{process_obj_defs}
\alias{process_obj_defs}
\title{Ensure all identifiers are set up correctly}
\usage{
process_obj_defs(obj_defs)
}
\arguments{
\item{obj_defs}{list of nodes' definitions (list of lists)}
}
\description{
Ensure all identifiers are set up correctly
}
