% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAS.R
\name{read.BAS}
\alias{read.BAS}
\title{Read basic (BAS) package file}
\usage{
read.BAS(filename, dis)
}
\arguments{
\item{filename}{character string;
file to be read}

\item{dis}{character string or object of class DIS.MFpackage;
a DIS package file (as file path or object), from which model dimensions
 will be read}
}
\value{
object of class BAS.MFpackage:\cr
  \code{$Options}: chr \code{[]}; options such as SHOWPROGRESS, FREE\cr
  \code{$IBOUND} : num \code{[NCOL, NROW, NLAY]} or \code{[NLAY]}; flags for
   inactive (0), active (1) and fixed head cells (-1)\cr
  \code{$HNOFLO} : num \code{[1]}; head assigned to no flow cells\cr
  \code{$STRT}   : num \code{[NCOL, NROW, NLAY]} or \code{[NLAY]}; initial
   heads\cr
}
\description{
Reads information from a MODFLOW-2000 BAS package file.  The BAS
 package contains information about the active region, the head assigned
 to no-flow cells and starting heads.
}
\examples{

}
