% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHB.R
\name{read.GHB}
\alias{read.GHB}
\title{Read general head boundary (GHB) package file}
\usage{
read.GHB(filename, nSP, FREE = TRUE)
}
\arguments{
\item{filename}{character string;
the package file to read}

\item{nSP}{integer \code{[1]} or DIS.MFpackage object;
number of stress periods or corresponding DIS package from which this
 can be read}

\item{FREE}{logical \code{[1]};
if TRUE, read assuming free format; normally FREE = TRUE is okay even if
 the file is in fixed format, providing entries have a space between
 them}
}
\value{
object of class GHB.MFpackage:\cr
  \code{$header} (1-row data.frame):\cr
    \code{..$MXACTB} (int): the maximum number of GHB boundary cells in any
      stress period\cr
    \code{..$IGHBCB} (int): unit number to which to save flow to GHB cells\cr
  \code{$spheaders} (data.frame with \code{nSP} rows):\cr
    \code{..$ITMP} (int): number of active GHB cells in each stress period
     or, if \eqn{<1}, signals to reuse information from previous stress
     period\cr
    \code{..$NP} (int): number of parameters active in each stress period (not
      that parameter-defined boundaries are not currently supported in
      \code{read.GHB})\cr
  \code{$data} (data.table):\cr
    \code{..$L} (int): layer\cr
    \code{..$R} (int): row\cr
    \code{..$C} (int): column\cr
    \code{..$BHead} (double): head at boundary above model datum
     (\eqn{h_{GHB}})\cr
    \code{..$Cond} (double): conductance of boundary (units
     \eqn{length^2/time})\cr
    \code{..$...}: any auxiliary variables that are included in the file
     but which don't play a role in the MODFLOW simulation
}
\description{
Reads information from a MODFLOW GHB package file.  The GHB package
 specifies the simplest kind of head dependent flux boundary condition
 that includes a conductance term.  The flux is calculated by:\deqn{
   Q_{GHB} = \frac{h_{model} - h_{GHB}}{Cond}
 }
}
\examples{

}
