% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFB.R
\name{read.HFB}
\alias{read.HFB}
\title{Read a horizontal flow barrier (HFB) package file}
\usage{
read.HFB(filename)
}
\arguments{
\item{filename}{character string}
}
\value{
list with class HFB.MFpackage, with elements:\cr
\code{$header}, 1-row data.frame:\cr
\code{..$NPHFB} int: number of HFBs specified as parameters (note that
 this feature is not yet supported by \code{read.HFB})\cr
\code{..$MXFB} int: the maximum number of flow barriers that will be
 defined by parameters (again, not yet supported)\cr
\code{..$NHFBNP} int: the number of barriers not defined by paramters\cr
\code{$data}, data.table:\cr
\code{..$L} int: layer\cr
\code{..$IROW1} int: row on side 1 of the barrier\cr
\code{..$ICOL1} int: column on side 1 of the barrier\cr
\code{..$IROW2} int: row on side 2 of the barrier\cr
\code{..$ICOL2} int: column on side 2 of the barrier\cr
\code{..$Hydchr} num: Hydraulic characteristic of the barrier, which is
 the hydraulic conductivity divided by the width
}
\description{
Reads information from a MODFLOW-2000 HFB6 package file.
 The HFB package places hydraulic barriers of set conductance between
 specified horizontally adjacent cells and is often used to represent
 artificial confining walls or geological faults.
}
