% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIV.R
\name{read.RIV}
\alias{read.RIV}
\title{Read river (RIV) package file}
\usage{
read.RIV(filename, nSP, FREE = TRUE)
}
\arguments{
\item{filename}{character string;
the package file to read}

\item{nSP}{integer \code{[1]} or DIS.MFpackage object;
number of stress periods or corresponding DIS package from which this
 can be read}

\item{FREE}{logical \code{[1]};
if TRUE, read assuming free format; normally FREE = TRUE is okay even if
 the file is in fixed format, providing entries have a space between
 them}
}
\value{
object of class RIV.MFpackage:\cr
  \code{$header} (1-row data.frame):\cr
    \code{..$MXACTB} (int): the maximum number of RIV boundary cells in any
      stress period\cr
    \code{..$IRIVCB} (int): unit number to which to save flow to RIV cells\cr
  \code{$spheaders} (data.frame with \code{nSP} rows):\cr
    \code{..$ITMP} (int): number of active RIV cells in each stress period
     or, if \eqn{<1}, signals to reuse information from previous stress
     period\cr
    \code{..$NP} (int): number of parameters active in each stress period (not
      that parameter-defined boundaries are not currently supported in
      \code{read.RIV})\cr
  \code{$data} (data.table):\cr
    \code{..$L} (int): layer\cr
    \code{..$R} (int): row\cr
    \code{..$C} (int): column\cr
    \code{..$Stage} (double): head at boundary above model datum
     (\eqn{h_{RIV}})\cr
    \code{..$Cond} (double): conductance of boundary (units
     \eqn{length^2/time})\cr
    \code{..$Rbot} (double): river bottom elevation; the difference
     between \code{Stage} and \code{Rbot} determines the maximum drainage
     of the river into the aquifer, although there is no limit to flow in
     the other direction\cr
    \code{..$...}: any auxiliary variables that are included in the file
     but which don't play a role in the MODFLOW simulation
}
\description{
Reads information from a MODFLOW RIV package file.  The RIV package
 specifies a head dependent flux boundary condition that includes a
 conductance term.  The flux is calculated by:\deqn{
   Q_{RIV} = \frac{h_{mb} - h_{RIV}}{Cond}
 }\cr
 where \eqn{h_{mb}} is the river bottom elevation or the model head for
 the cell, whichever is higher
}
\examples{
fnms <- system.file(c("rflow_mf_demo.dis",
                      "rflow_mf_demo.riv"), package = "Rflow")

# get model information from DIS package file
dis <- read.DIS(fnms[1L])

riv <- read.RIV(fnms[2L], dis)
# or, if you already know how many stress periods there are
riv <- read.RIV(fnms[2L], 15L)

class(riv)
str(riv)

}

