% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR.R
\name{read.STR}
\alias{read.STR}
\title{Read stream (STR) package file}
\usage{
read.STR(filename, nSP)
}
\arguments{
\item{filename}{character string;
the package file to read}

\item{nSP}{integer \code{[1]} or DIS.MFpackage object;
number of stress periods or corresponding DIS package from which this
 can be read}
}
\value{
list with class STR.MFpackage:\cr
\code{$header} 1-row data.frame:\cr
  \code{..$MXSTRM} (int): maximum number of stream reaches\cr
  \code{..$NSS} (int): maximum number of segments\cr
  \code{..$NTRIB} (int): maximum number of tributary segments that join
   one downstream segment\cr
  \code{..$NDIV} (int): flag, which when positive, specifies that
   diversions from segments are to be simulated\cr
  \code{..$ICALC} (int): flag, which when positive, specifies that
   stream stages in reaches are to be calculated\cr
  \code{..$CONST} (num): Constant value used in calculating stream stage in
   reaches. It is specified whenever ICALC is greater than zero.  A unit
   conversion between stream flow units and the units used for the MODFLOW
   simulation.\cr
  \code{..$ISTCB1} (int): unit number to which to save stream-aquifer
   exchange\cr
  \code{..$ISTCB2} (int): unit number to which to save stream flow \cr
\code{$spheaders} data.frame with 1 row for each stress period:\cr
  \code{..$ITMP} (int): if < 0, read stream data from the previous
   stress period, otherwise it is the number of reaches to read for the
   current stress period\cr
  \code{..$IRDFLG} (int): if > 0, don't print input data for this sp\cr
  \code{..$IPTFLG} (int): if > 0, don't print results for this sp\cr
\code{$data}\cr
  \code{..$L} (int): layer\cr
  \code{..$R} (int): row\cr
  \code{..$C} (int): column\cr
  \code{..$Seg} (int): number assigned to a group of reaches\cr
  \code{..$Reach} (int): sequential number in a segment that begins with one
   for the farthest upstream reach and continues in downstream order to the
   last reach in the segment\cr
  \code{..$Flow} (num): streamflow, in length cubed per time, entering
   a segment\cr
  \code{..$Stage} (num): stream stage, in units of length\cr
  \code{..$Cond} (num): streambed hydraulic conductance\cr
  \code{..$Sbot} (num): elevation of the bottom of the streambed\cr
  \code{..$Stop} (num): elevation of the top of the streambed\cr
  if \code{ICALC > 0}\{\cr
  \code{..$Width} (num): width of the stream channel\cr
  \code{..$Slope} (num): slope of the stream channel in each reach\cr
  \code{..$Rough} (num): Manning's roughness coefficient for each stream
   reach\cr
  \}\cr
  if \code{NTRIB > 0}\{\cr
  \code{..$ITrib1} (int): number of the first tributary segment\cr
  ...\cr
  \code{..$ITrib<NTrib>} (int):\cr
  \}\cr
  if \code{NDIV > 0}\{\cr
  \code{..$Iupseg} (int): number of the upstream segment from which
   water is diverted\cr
  \}
}
\description{
Reads information from a MODFLOW STR package file.  The stream package
 routes stream flow through connected cells and simultaneously
 calculates water exchange between streams and the aquifer.
}
\references{
Prudic, D.E., 1989. Documentation of a Computer Program to Simulate Stream-Aquifer Relations Using a Modular, Finite-Difference Ground-water Flow Model, Carson City, Nevada. Available at: https://pubs.er.usgs.gov/publication/ofr88729.
}

