% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEL.R
\name{read.WEL}
\alias{read.WEL}
\title{Read well (WEL) package file}
\usage{
read.WEL(filename, nSP, dis, MF2k = TRUE)
}
\arguments{
\item{filename}{character string;
WEL package file}

\item{nSP}{integer \code{[1]};
number of stress periods}

\item{dis}{DIS.MFpackage object;
if \code{nSP} is not given, it can be read from \code{nrow(dis$sps)},
 but not necessary if \code{nSP} is given}

\item{MF2k}{logical \code{[1]};
is this a MODFLOW 2000+ WEL package?}
}
\value{
list \code{[2]} of class WEL.MFpackage, with elements:
\code{$header}:
  \code{$MXACTW} the maximum number of active wells in a stress period
  \code{$IWELCB} unit number to which well flux array is saved (if > 0)
\code{$spheaders}:
headers for each stress period, a data.frame with columns:
  \code{$ITMP} (int): number of active wells in stress period, or (if < 0)
  indicates to read from previous stress period
  \code{$NP} (int): number of parameters to read for this stress period
  (always 0 if MF2k = FALSE)
\code{$read}: logical vector of length nSP indicating for which stress periods
well information is to be read
\code{$data}:
a data.table with the following columns
  \code{$sp} (int): stress period (key)
  \code{$C} (int): column
  \code{$R} (int): row
  \code{$L} (int): layer
  \code{$Q} (num): flux (negative indicates abstraction, positive indicates
  injection)
}
\description{
Reads information from a MODFLOW WEL package file.  The WEL package
 assigns fixed sink/ source fluxes to individual cells and as such is
 particularly appropriate for representing wells of known abstraction or
 injection rate.
}
\examples{
fnms <- system.file(c("rflow_mf_demo.dis",
                      "rflow_mf_demo.wel"), package = "Rflow")

# get model information from DIS package file
dis <- read.DIS(fnms[1L])

wel <- read.WEL(fnms[2L], dis = dis)
# or, if you already know how many stress periods there are
wel <- read.WEL(fnms[2L], 15L)

class(wel)
str(wel)

}
