% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODoutdata.R
\name{readCBB.arr}
\alias{readCBB.arr}
\title{Read Cell-By-Cell Budget Array}
\usage{
readCBB.arr(file, flux.bn = 4L, show.help = TRUE, nf.to.NA = F, hds,
  CRLs = list("all", "all", "all"), sp_ts = "all", artys = "all")
}
\arguments{
\item{file}{character string;
file name to read}

\item{flux.bn}{integer \code{[1]};
number of bytes occupied by a flow array element (generally 4)}

\item{show.help}{logical \code{[1]};
display progress markers and help message at the end?}

\item{nf.to.NA}{logical \code{[1]};
make cells for no-flow cells into NA?  (requires a head array to be
 given to \code{hds}; can be slow)}

\item{hds}{list, as would be output from \code{\link{readHDS.arr}}}

\item{CRLs}{list \code{[3]};
each element should either by \code{"all"} or a subset of columns (first
 element), rows (second element) or layers (third element) to return a
 sub-array}

\item{sp_ts}{character strings \code{[]} or integer matrix \code{[, 2]};
either \code{"all"} or a set of <stress period _ time step>s to read}

\item{artys}{character strings \code{[]};
\code{"all"} or output array types to read (e.g. "FlowRightFace",
 "RiverLeakage")}
}
\value{
5D numeric array \code{[NCOL, NROW, NLAY, NTS, NARTY]}.  The fourth and
 fifth dimensions are given appropriately formatted dimnames, such that
 specific output types may be subsetted by name.
}
\description{
Read Cell-By-Cell Budget Array
}

