% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODoutdata.R
\name{readUCN.arr}
\alias{readUCN.arr}
\title{Read Unformatted Concentration File}
\usage{
readUCN.arr(file, time.only = F, time.bn = 4L, conc.bn = 4L,
  show.help = T, nf.to.NA = T, nf.val = -1, lays = "all",
  sp_ts_tts = "all")
}
\arguments{
\item{file}{character string}

\item{time.only}{logical \code{[1]};
if \code{TRUE}, then only return the time at the end of each time step}

\item{time.bn}{integer \code{[1]};
number of bytes in a time record (generally 4)}

\item{conc.bn}{integer \code{[1]};
number of bytes in an array record (generally 4)}

\item{show.help}{logical \code{[1]};
show progress markers and explanatory notes?}

\item{nf.to.NA}{logical \code{[1]};
convert inactive cells to NA?}

\item{nf.val}{numeric \code{[1]};
value denoting inactive cells (commonly -1)}

\item{lays}{\code{"all"} or integer \code{[]};
return all layers, or else a subset}

\item{sp_ts_tts}{\code{"all"}, character strings \code{[]} or integer matrix \code{[, 3]};
return all transport time steps, or a subset of
 <stress period_time step_transport time step>s}
}
\value{
either a list with two elements (with \code{time.only = FALSE}):\cr
  \code{$Concentration}: num \code{[NCOL, NROW, NLAY, NTTS]}\cr
  \code{$time}: num \code{[NTTS]}; time at end of transport time steps,
   relative to model start\cr
or just the time values
}
\description{
Read Unformatted Concentration File
}
