% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{rf}
\alias{rf}
\title{Implicit cache of a function and processing of the given call.}
\usage{
rf(fn_call, rflow_options = get_rflow_options())
}
\arguments{
\item{fn_call}{Function call to be processed.}

\item{rflow_options}{List of options created using \code{get_rflow_options}.}
}
\value{
The cached version of the function.
}
\description{
Implicit cache of a function and processing of the given call.
}
