% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{rflow Pipeline}
\alias{rflow Pipeline}
\alias{rflowPipeline}
\title{rflow Pipeline}
\usage{
rflowPipeline(eval_metric, test_percent, ...)
}
\arguments{
\item{eval_metric}{Evaluation metric to be used, e.g. accuracy_score, mean_squared_error. A list can be found here: http://scikit-learn.org/stable/modules/classes.html#sklearn-metrics-metrics}

\item{...}{Additional transformations can be passed here, such as custom model. See demos}
}
\description{
Main function used to run TensorFlow in pipeline. See demos for usage
}

