% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{rflow}
\alias{rflow}
\title{Initialize a new DAG}
\usage{
rflow(
  path = NULL,
  cache = TRUE,
  persistence = TRUE,
  logging = if (length(path)) TRUE else FALSE
)
}
\arguments{
\item{path}{path to rflow home folder. Nodes' definitions are searched for in this folder, cache and other data is saved into...}

\item{persistence}{logical;}

\item{logging}{logical;}

\item{caching}{logical;}
}
\value{
A new rflow object.
}
\description{
Initialize a new DAG
}
\examples{
\dontrun{
RF <- new_rflow()
}
}
