% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rflow_check_acyclic}
\alias{rflow_check_acyclic}
\title{Detect cycles explicitly (used for early validation with a friendlier message)}
\usage{
rflow_check_acyclic(task_ids, upstream_map)
}
\description{
Detect cycles explicitly (used for early validation with a friendlier message)
}
\keyword{internal}
