% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{rflow_code}
\alias{rflow_code}
\title{Generate R code from nodes}
\usage{
rflow_code(outputVar = "", operator = c("<-", "\%>\%", "+")[1])
}
\arguments{
\item{outputVar}{Character scalar with name of the variable to capture the output}

\item{operator}{Character scalar giving an operator to compose the function calls}
}
\value{
Inserts generated code as text and returns NULL invisibly
}
\description{
Takes a connection to the node app and inserts R code 
  corresponding to nodes
}

