% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rflow_db_runs}
\alias{rflow_db_runs}
\title{Fetch dag_run history from the metadata database}
\usage{
rflow_db_runs(con, dag_id = NULL)
}
\arguments{
\item{con}{connection from [rflow_db_connect()]}

\item{dag_id}{optional filter}
}
\description{
Fetch dag_run history from the metadata database
}
