% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rflow_db_task_instances}
\alias{rflow_db_task_instances}
\title{Fetch task_instance rows for a given run}
\usage{
rflow_db_task_instances(con, run_id)
}
\arguments{
\item{con}{connection from [rflow_db_connect()]}

\item{run_id}{the dag run id}
}
\description{
Fetch task_instance rows for a given run
}
