% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{rflow_eval}
\alias{rflow_eval}
\title{Evaluate R code from nodes}
\usage{
rflow_eval(operator = c("<-", "\%>\%", "+")[1], flows = "")
}
\arguments{
\item{tcp_msg}{Character scalar holding a message from the tcp server in JSON format}
}
\value{
Final return value of the generated code
}
\description{
Takes a connection to the node app and evaluates R code 
  corresponding to nodes
}

