% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{rflow_pkgnodes}
\alias{rflow_pkgnodes}
\title{Generate a JSON Representation of User-facing Functions}
\usage{
rflow_pkgnodes(pkg)
}
\arguments{
\item{pkg}{Character scalar naming a package}
}
\value{
Translate package functions to nodes and return NULL invisibly
}
\description{
Generates a JSON with function names, arguments, default values, and
  manuals for all user-facing functions in a package
}

