% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{rflow_receive}
\alias{rflow_receive}
\title{JSON Reader}
\usage{
rflow_receive(con)
}
\arguments{
\item{con}{Connection object point to node application's tcp server}
}
\value{
Character scalar holding a message in JSON format
}
\description{
Receive a JSON Representation of Generated Functions from the NodeRed App
}

