% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{rflow_start}
\alias{rflow_start}
\title{Start RFlow}
\usage{
rflow_start(viewer = TRUE, comm_port = "1338")
}
\arguments{
\item{viewer}{Logical scalar determining whether to open the node-red environment
in the Rstudio viewer or a browser window}

\item{comm_port}{Character scalar giving communication server port}

\item{node_port}{Character scalar giving node-red environment port}
}
\value{
Connection object for the communication server
}
\description{
Takes app and communication ports to start
}

