% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rflow_topo_sort}
\alias{rflow_topo_sort}
\title{Topologically sort a set of tasks given an adjacency list of dependencies}
\usage{
rflow_topo_sort(task_ids, upstream_map)
}
\arguments{
\item{task_ids}{character vector of all task ids}

\item{upstream_map}{named list: task_id -> character vector of upstream task_ids}
}
\value{
list(order = character vector in valid execution order, layers = list of character vectors,
  where layers[[i]] are tasks that can run in parallel at "depth" i)
}
\description{
Topologically sort a set of tasks given an adjacency list of dependencies
}
\keyword{internal}
