% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{RIARRAY}
\alias{RIARRAY}
\title{Formatted array}
\usage{
RIARRAY(arr = NULL, df = NULL, CNSTNT = 0, FMTIN_type = "e",
  FMTIN_w = 12, FMTIN_d = 4, FMTIN_N = ncol(arr), IPRN = -1,
  header = T, nrowmod, ncolmod, bgC = 0, flag.no = NA)
}
\arguments{
\item{flag.no}{}
}
\value{
character string
}
\description{
Formatted array
}
\examples{
Rflow:::RIARRAY(matrix(1:4, 2, 2), FMTIN_type = "I", FMTIN_w = 3L)
Rflow:::RIARRAY(matrix(1:4, 2, 2), FMTIN_type = "e", FMTIN_w = 12L)
Rflow:::RIARRAY(CNSTNT = 1L, FMTIN_type = "I", FMTIN_w = 3L)

}

