% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{RIARRAY.splitlayers}
\alias{RIARRAY.splitlayers}
\title{Formatted array, split by layer}
\usage{
RIARRAY.splitlayers(NLAY, CNSTNT = rep(0, NLAY), arr = NULL,
  compress = TRUE, collapse = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
character string
}
\description{
Formatted array, split by layer
}
