% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{rmnewline}
\alias{rmnewline}
\title{Remove a tailing line end}
\usage{
rmnewline(s)
}
\arguments{
\item{s}{vector of character strings}
}
\value{
vector of character strings with any \code{"\\n"}s removed from the ends
}
\description{
Remove a tailing line end
}
