% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{send_nodes}
\alias{send_nodes}
\title{JSON Writer}
\usage{
send_nodes(tcp_msg, con)
}
\arguments{
\item{con}{Connection object point to node application's tcp server}

\item{pkg_nodes}{Character scalar holding JSON representation of nodes}
}
\value{
NULL invisibly
}
\description{
Send a JSON Representation of User-facing Functions to the NodeRed App
}

