% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{set_current_eddy}
\alias{set_current_eddy}
\title{Set the current eddy to be used in future rflow calls.}
\usage{
set_current_eddy(eddy_name, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to allow retrieving later.}

\item{eddy_env}{An environment where to put (bind) the eddy.}
}
\value{
An eddy object to be used for storing rflows.
}
\description{
Set the current eddy to be used in future rflow calls.
}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_current_eddy}}, \code{\link{get_eddy}},
  \code{\link{new_eddy}}
}
\concept{eddy functions}
