% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{show_run}
\alias{show_run}
\title{Print a Gantt-style text summary of a single DAG run's task instances}
\usage{
show_run(con, run_id)
}
\arguments{
\item{con}{connection from [rflow_db_connect()]}

\item{run_id}{the run to summarize}
}
\description{
Print a Gantt-style text summary of a single DAG run's task instances
}
