% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{skiparr}
\alias{skiparr}
\title{Skip over unneed binary arrays}
\usage{
skiparr(con, nvec, bnvec)
}
\arguments{
\item{con}{file connection}

\item{nvec}{integer \code{[]};
numbers of entries}

\item{bnvec}{integer \code{[]};
numbers of bytes per entry}
}
\value{
\code{NULL}
}
\description{
Skip over unneed binary arrays
}

