% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitute...}
\alias{substitute...}
\title{Substitute ... for a list of un-evaluated expressions}
\usage{
substitute...()
}
\value{
a list of all arguments matched by ...
}
\description{
Substitute ... for a list of un-evaluated expressions
}
