% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_model_helpers.R
\name{Tensor Activator}
\alias{Tensor Activator}
\alias{TensorActivator}
\title{Tensor Activator}
\usage{
TensorActivator(NAME, ...)
}
\arguments{
\item{NAME}{The name of the activation method to be used. Available activators are: linear_regression, logistic_regression, etc}

\item{...}{Additional argument except the tensor input}
}
\description{
This function performs activation function on the tensor
}

