% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_model_helpers.R
\name{Tensor Operator}
\alias{Tensor Operator}
\alias{TensorOperator}
\title{Tensor Operater}
\usage{
TensorOperator(NAME, ...)
}
\arguments{
\item{NAME}{The name of the operation method to be used. Available operators are: conv2d, dnn}

\item{...}{Additional argument except the tensor input}
}
\description{
This function performs operations on tensor input, such as conv2d for convolutional layer
and dnn for deep neural network layer
}

