% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_model_helpers.R
\name{Tensor Processor Transformation}
\alias{Tensor Processor Transformation}
\alias{TensorProcessor.transform}
\title{Tensor Processor Transformation}
\usage{
TensorProcessor.transform(dataType = "train")
}
\arguments{
\item{dataType}{Whether this is applying transformation for 'train' or 'test'}
}
\description{
This function fit and transform training text OR transform testing text
}

