% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_model_helpers.R
\name{Tensor Processor}
\alias{Tensor Processor}
\alias{TensorProcessor}
\title{Tensor Processor}
\usage{
TensorProcessor(NAME, ...)
}
\arguments{
\item{NAME}{The name of processing function. A list can be found here: https://github.com/google/skflow/blob/master/skflow/preprocessing/text.py}

\item{...}{Additional paramerters passed to processing function}
}
\description{
This function performs processing function on the tensor
}

