% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_model_helpers.R
\name{Tensor Transformer}
\alias{Tensor Transformer}
\alias{TensorTransformer}
\title{Tensor Transformaer}
\usage{
TensorTransformer(NAME, ...)
}
\arguments{
\item{NAME}{The name of the transoformation method to be used.}

\item{...}{Additional argument except the tensor input
# Available transformation method are list here: http://www.tensorflow.org/api_docs/python/array_ops.html#tensor-transformations}
}
\description{
This function performs transformation on tensor input and returns tensor input
}

