% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{TensorFlow Custom Model Builder}
\alias{TensorFlow Custom Model Builder}
\alias{TensorFlowEstimator}
\title{TensorFlow Custom Model Builder}
\usage{
TensorFlowEstimator(n_classes, ...)
}
\arguments{
\item{n_classes}{The number of classes for the target variable}

\item{...}{Additional parameters including batch_size, steps, learning_rate, etc. A list can be found here: https://github.com/google/skflow/blob/master/skflow/__init__.py#L29}
}
\description{
This function is used when a custom model is written. 
Any custom model will be passed automatically into this function.
}

