% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{TensorFlow DNN Classifier}
\alias{TensorFlow DNN Classifier}
\alias{TensorFlowDNNClassifier}
\title{TensorFlow DNN Classifier}
\usage{
TensorFlowDNNClassifier(hidden_units, n_classes, ...)
}
\arguments{
\item{hidden_units}{A vector of the number of hidden units in each layer}

\item{n_classes}{The number of classes in the target}

\item{...}{Additional parameters such as keep_prob, a list can be found here: https://github.com/google/skflow/blob/master/skflow/ops/dnn_ops.py}
}
\description{
Deep Neural Network classifier
}

