% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_experiment.R
\name{TensorFlow DNN Regressor}
\alias{TensorFlow DNN Regressor}
\alias{TensorFlowDNNRegressor}
\title{TensorFlow DNN Regressor}
\usage{
TensorFlowDNNRegressor(hidden_units, ...)
}
\arguments{
\item{hidden_units}{A vector of the number of hidden units in each layer}

\item{...}{Additional parameters such as keep_prob. A list can be found here: https://github.com/google/skflow/blob/master/skflow/ops/dnn_ops.py}
}
\description{
Deep Neural Network regressor
}

