% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{to_ns}
\alias{to_ns}
\title{Assigns a value to a name space (an environment or a reactiveValues).}
\usage{
to_ns(x, var, ns)
}
\arguments{
\item{x}{Value to assign.}

\item{var}{The name (as string) of the variable.}

\item{ns}{The name space, either an \code{environment} or a 
\code{Shiny reactiveValues} object.}
}
\value{
The initial value, \code{x}
}
\description{
Assigns a value to a name space (an environment or a reactiveValues).
}
