% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{trigger_manual}
\alias{trigger_manual}
\alias{trigger_manual.node}
\alias{trigger_manual.character}
\alias{trigger_manual.rflow}
\title{Obtain value/object represented by a node}
\usage{
trigger_manual(x, ...)

\method{trigger_manual}{node}(x)

\method{trigger_manual}{character}(x, rflow)

\method{trigger_manual}{rflow}(rflow, x)
}
\arguments{
\item{x}{a node or node's id}

\item{rflow}{rflow object}
}
\value{
Value/object represented by given node.
}
\description{
Obtain value/object represented by a node
}
