% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{update_nodes}
\alias{update_nodes}
\title{Batch update nodes}
\usage{
update_nodes(objs, rflow, verbose = FALSE)
}
\arguments{
\item{objs}{list of nodes' definitions}

\item{rflow}{an rflow object}

\item{verbose}{logical; print verbose output?}
}
\description{
Batch update nodes
}
