% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{update_obj}
\alias{update_obj}
\title{update object (call whenever definition in config file could have potentially changed)}
\usage{
update_obj(obj, rflow, verbose = FALSE)
}
\description{
update object (call whenever definition in config file could have potentially changed)
}
