% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_deps.R
\name{verify_dependencies}
\alias{verify_dependencies}
\title{Verify consistency of user-declared dependencies with node's behaviour}
\usage{
verify_dependencies(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Verify consistency of user-declared dependencies with node's behaviour
}
