% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichnc.R
\name{which.nc}
\alias{which.nc}
\title{"which" for NetCDF arrays}
\usage{
which.nc(ncfile, variable, FUN = Negate(is.na), ..., arr.ind = FALSE,
  useNames = TRUE, size.threshold = 1e+07)
}
\arguments{
\item{ncfile}{NetCDF connection object}

\item{variable}{variable name within NetCDF data set}

\item{FUN}{function to operate on array, which should produce a logical
vector or array}

\item{...}{additional arguments to FUN}

\item{arr.ind, useNames}{passed to which}

\item{size.threshold}{how large can the array be before the function
performs on chunks?}
}
\value{
integer vector (arr.ind = FALSE) or matrix (arr.ind = TRUE), as with
`which`
}
\description{
A version of "which" for NetCDF datasets that does not require prior
loading of arrays and avoids loading the whole of large arrays at once.
}
\examples{
library(RNetCDF)

mfdata <- open.nc(system.file("rflow_mf_demo.nc", package = "Rflow"))

# find which cells have a well in, by C, R, L reference
unique(which.nc(mfdata, "Wells", `!=`, 0, arr.ind = TRUE)[, 1:3],
       MARGIN = 1L)

}
