% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAS.R
\name{write.BAS}
\alias{write.BAS}
\title{Write MODFLOW BAS package file}
\usage{
write.BAS(BAS, filename, unit = 1L, title = "", MF2k = TRUE, sps, NCOL,
  NROW, NLAY)
}
\arguments{
\item{BAS}{object of class BAS.MFpackage, as would be returned by
\code{\link{read.BAS}}}

\item{filename}{character string;
file to write to, ideally ending in \code{".bas"}}

\item{unit}{integer \code{[1]};
unit number that the BAS file will be assigned to}

\item{title}{character string;
optional title to prefix to the file, for your reference (not used if MF2k =
 FALSE)}

\item{MF2k}{logical;
should the file be for MODFLOW 2000, or MODFLOW 96 (in which case stress
 period information is written from sps)}

\item{sps}{data.frame or object of class DIS.MFpackage;
only required if \code{MF2k = FALSE}; a data frame of stress period
 information with the same columns as described in \code{\link{read.DIS}},
 except that the \code{TR} column is not required (ignored if provided)}

\item{NCOL, NROW, NLAY}{integer \code{[1]}; only required if \code{MF2k = FALSE} and
\code{class(sps) != "DIS.MFpackage"}}
}
\value{
\code{NULL}
}
\description{
Writes information from a BAS.MFpackage list object to a MODFLOW-readable
 BAS package file.
}
\examples{

}
