% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHB.R
\name{write.GHB}
\alias{write.GHB}
\title{Write a general head boundary (GHB) package file}
\usage{
write.GHB(GHB, filename, title, MF2k = TRUE, write.auxiliaries = TRUE)
}
\arguments{
\item{GHB}{GHB.MFpackage object, as would be returned by \code{\link{read.GHB}}}

\item{filename}{character string;
file to write to}

\item{title}{character string;
optional title to prefix to the file, for your reference (not used if
 \code{MF2k = FALSE})}

\item{MF2k}{logical \code{[1]};
should the file be readable by MODFLOW 2000 and higher, or
 (\code{FALSE}) keep back-compatibility with MF96; in the case of the
 GHB package, this means no title lines}

\item{write.auxiliaries}{logical [1];
whether to include any auxiliary parameters in \code{RIV} in the package
 file}
}
\description{
Writes information from a GHB.MFpackage list object to a MODFLOW-readable
 GHB package file.
}
\examples{

}

