% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIV.R
\name{write.RIV}
\alias{write.RIV}
\title{Write a river (RIV) package file}
\usage{
write.RIV(RIV, filename, title, MF2k = TRUE, write.auxiliaries = TRUE)
}
\arguments{
\item{RIV}{RIV.MFpackage object, as would be returned by \code{\link{read.RIV}}}

\item{filename}{character string;
file to write to}

\item{title}{character string;
optional title to prefix to the file, for your reference (not used if
 \code{MF2k = FALSE})}

\item{MF2k}{logical \code{[1]};
should the file be readable by MODFLOW 2000 and higher, or
 (\code{FALSE}) keep back-compatibility with MF96; in the case of the
 RIV package, this means no title lines}

\item{write.auxiliaries}{logical [1];
whether to include any auxiliary parameters in \code{RIV} in the package
 file}
}
\description{
Writes information from a RIV.MFpackage list object to a MODFLOW-readable
 RIV package file.
}
\examples{

fnm <- system.file("rflow_mf_demo.riv", package = "Rflow")
riv <- read.RIV(fnm, 15L, FALSE)

# write modified riv package to file
write.RIV(riv, "RFLOW_EXAMPLE.riv", "example")

}

