% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR.R
\name{write.STR}
\alias{write.STR}
\title{Write a stream (STR) package file}
\usage{
write.STR(STR, filename)
}
\arguments{
\item{STR}{STR.MFpackage object}

\item{filename}{character string}
}
\description{
Writes information from a STR.MFpackage list object to a MODFLOW-readable
 STR package file.
}
