% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEL.R
\name{write.WEL}
\alias{write.WEL}
\title{Write a well (WEL) package file}
\usage{
write.WEL(WEL, filename, title = "", MF2k = TRUE)
}
\arguments{
\item{WEL}{object of class WEL.MFpackage}

\item{filename}{character string;
file to which package information is to be written}

\item{title}{character string;
optional title to prefix to the file, for your reference (not used if MF2k =
 FALSE)}

\item{MF2k}{}
}
\description{
Writes information from a WEL.MFpackage list object to a MODFLOW-readable
 WEL package file.
}
\examples{
# read WEL package
fnm <- system.file("rflow_mf_demo.wel", package = "Rflow")
wel <- read.WEL(fnm, 15L)

# meaningful modification e.g. add a stress period with no abstraction
wel$spheader <- rbind(wel$spheader, list(0L, 0L))
wel$read <- c(wel$read, TRUE)

# write modified wel package to file
write.WEL(wel, "RFLOW_EXAMPLE.wel", "example: added stress period")

}
