% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{write_nodes}
\alias{write_nodes}
\title{Save a JSON Representation of User-facing Functions to Disk}
\usage{
write_nodes(pkg_nodes, file_name, path)
}
\arguments{
\item{pkg_nodes}{Character scalar holding JSON representation of nodes}

\item{file_name}{Character scalar in name.json format}

\item{path}{Character scalar giving write location}
}
\value{
JSON file on disc and return NULL invisibly
}
\description{
Writes node specification to disk
}

