% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{cache_memory}
\alias{cache_memory}
\title{Get the memory only cache engine.}
\usage{
cache_memory()
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
All cache data will be stored only in memory.
}
\examples{
cache_in_memory <- cache_memory()

}
\seealso{
Other cache functions: \code{\link{cache_file}},
  \code{\link{cache_memory_file}},
  \code{\link{default_cache}}
}
\concept{cache functions}
