% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{collect}
\alias{collect}
\alias{collect.R6Flow}
\alias{collect.R6FlowElement}
\title{Get the data from an \code{R6Flow} or an \code{R6FlowElement} object.}
\usage{
\method{collect}{R6Flow}(x, ...)

\method{collect}{R6FlowElement}(x, ...)
}
\arguments{
\item{x}{Function cached with RFlow.}

\item{...}{Element of the output data to be selected. If present, it should
be named \code{name}, otherwise the first item of the \code{...} list
will be used. The default is \code{name = NULL}, which returns all the
data.}
}
\value{
Data associated with the output of the function.
}
\description{
Get the data from an \code{R6Flow} or an \code{R6FlowElement} object.
}
