% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{collect}
\alias{collect}
\title{Get data from an RFlow object.}
\usage{
collect(rf_fn, what = NULL)
}
\arguments{
\item{rf_fn}{Function cached with RFlow.}

\item{what}{Element of the output data to be selected.}
}
\value{
Data associated with the output of the function.
}
\description{
Get data from an RFlow object.
}
