% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{default_cache}
\alias{default_cache}
\title{Get the default cache engine.}
\usage{
default_cache()
}
\value{
A cache object that inherits from \code{R6Cache}.
}
\description{
Currently, the default cache engine is \code{cache_memory()}, but this 
  might change in the future.
}
\examples{
\dontrun{
current_cache <- default_cache()
}

}
\seealso{
Other cache functions: \code{\link{cache_file}},
  \code{\link{cache_memory_file}},
  \code{\link{cache_memory}}
}
\concept{cache functions}
