% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{delete_eddy}
\alias{delete_eddy}
\title{Delete eddy and ALL its data from ALL cache layers.}
\usage{
delete_eddy(eddy_name, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to allow retrieving later.}

\item{eddy_env}{An environment where to put (bind) the eddy.}
}
\value{
Nothing (\code{NULL}).
}
\description{
Delete eddy and ALL its data from ALL cache layers.
}
\seealso{
Other eddy functions: \code{\link{get_current_eddy}},
  \code{\link{get_eddy}}, \code{\link{new_eddy}},
  \code{\link{set_current_eddy}}, \code{\link{use_eddy}}
}
\concept{eddy functions}
