% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{delete_eddy}
\alias{delete_eddy}
\title{Delete eddy and ALL its data from ALL cache layers (memory and/or disk).}
\usage{
delete_eddy(eddy_name, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name of the \code{eddy} to be deleted.}

\item{eddy_env}{The environment in which the eddy is put (bound).}
}
\value{
Nothing (\code{NULL}).
}
\description{
Delete eddy and ALL its data from ALL cache layers (memory and/or disk).
}
\examples{
new_eddy("eddy_new")
# processing steps ...
delete_eddy("eddy_new")

}
\seealso{
Other eddy functions: \code{\link{get_current_eddy}},
  \code{\link{get_eddy}}, \code{\link{new_eddy}},
  \code{\link{set_current_eddy}}, \code{\link{use_eddy}}
}
\concept{eddy functions}
