% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_deps.R
\name{detect_nodes}
\alias{detect_nodes}
\title{Detect dependencies}
\usage{
detect_nodes(x, rflow, found = c(), space = "", depth = 0L,
  depthmax = 20L, verbose = FALSE)
}
\arguments{
\item{space}{}
}
\description{
Detect dependencies
}
