% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_deps.R
\name{detect_nodes}
\alias{detect_nodes}
\title{Detect dependencies}
\usage{
detect_nodes(
  x,
  rflow,
  noderef = NULL,
  found = c(),
  space = "",
  depth = 0L,
  depthmax = 20L,
  verbose = FALSE,
  debugging = FALSE
)
}
\arguments{
\item{space}{character; prefix to be added when printing}

\item{depth}{DO NOT SET THIS ARG! (counter of the level how deep the recurent algo is currently entering)}

\item{depthmax}{limit on the depth to dive in}

\item{verbose}{logical; print output?}

\item{envir}{R environemnt}
}
\value{
detected nodes' id
}
\description{
Detect dependencies
}
