% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{element}
\alias{element}
\alias{[.R6Flow}
\title{Extract an element from an \code{R6Flow} object.}
\usage{
element(x, name = NULL)

\method{[}{R6Flow}(x, name)
}
\arguments{
\item{x}{Function cached with RFlow.}

\item{name}{Element of the output data to be selected. The default is 
\code{name = NULL}, which returns the element version of the \code{R6Flow} 
input object.}
}
\value{
An object with class \code{Element}.
}
\description{
Extract an element from an \code{R6Flow} object.
}
