% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{element}
\alias{element}
\alias{[.R6Flow}
\title{Extract an element from an \code{R6Flow} object.}
\usage{
element(flow, name = NULL)

\method{[}{R6Flow}(flow, name)
}
\arguments{
\item{flow}{A flow object, e.g. as returned by \code{\link{flow_fn}}.}

\item{name}{Element of the output data to be selected. The default is 
\code{name = NULL}, which returns the element version of the \code{R6Flow} 
input object.}
}
\value{
An object with class \code{Element}.
}
\description{
Extract an element from an \code{R6Flow} object.
}
\examples{
fn <- function(x, y) { list(x = x, y = y, Z = 6) }
flowed_fn <- flow_fn(2, 3, fn = fn) 
flow_element <- element(flowed_fn, "x")

fn <- function(x, y) { list(x = x, y = y, Z = 7) }
flowed_fn <- flow_fn(2, 3, fn = fn)
element_name <- flowed_fn["x"]$elem_name

}
