% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{flow}
\alias{flow}
\title{Implicit cache of a function and processing of the given call.}
\usage{
flow(fn_call, fn_id = NULL, flow_options = get_flow_options())
}
\arguments{
\item{fn_call}{Function call to be processed.}

\item{fn_id}{Optional id to uniquely identify the function. By default,
rflow functions reuse the cache if the same function is given. The id 
allows the user to suppress console messages and to
indicate whether to reuse the old cache or create a new one.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The flow object.
}
\description{
Implicit cache of a function and processing of the given call.
}
