% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{flow_fn}
\alias{flow_fn}
\title{Implicit cache of a function and of the given call.}
\usage{
flow_fn(..., fn = NULL, fn_id = NULL, flow_options = get_flow_options())
}
\arguments{
\item{...}{Named arguments to pass to \code{fn}.}

\item{fn}{The function to apply to the data frame. It must accept a data
frame as the first argument and a numeric index as the second argument.}

\item{fn_id}{Optional id to uniquely identify the function. By default,
rflow functions reuse the \code{cache} if the same function is given. The id 
allows the user to suppress console messages and to explicitly
indicate whether to reuse the old \code{cache} or create a new one.}

\item{flow_options}{List of options created using \code{get_flow_options}.}
}
\value{
The flow object.
}
\description{
Implicit cache of a function and of the given call.
}
\details{
Arguments \code{fn}, \code{fn_id} and \code{flow_options}, when provided,
  must be named. Argument \code{fn} must be always provided.
}
\examples{
fn <- function(x, y) { x + y + 2 }
flowed_fn <- flow_fn(2, 3, fn = fn)
collected_result <- flowed_fn \%>\%
    collect()
# usage with rflow pipes
fn2 <- function(x, y) { x * y }
collected_pipe_result <- flowed_fn \%>\%
    flow_fn(2, fn = fn2) \%>\%
    collect()

}
