% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{forget}
\alias{forget}
\title{Forgets the computation for the current state.}
\usage{
forget(flow, state = "current")
}
\arguments{
\item{flow}{An flow object, e.g. as returned by \code{\link{flow_call}}.}

\item{state}{An flow state. Only the \code{current} state is 
accepted for now.}
}
\value{
A logical value, whether the deletion was successful.
}
\description{
Forgets the computation for the current state.
}
