% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{forget}
\alias{forget}
\title{Forgets the computation for the current state.}
\usage{
forget(rflow)
}
\arguments{
\item{rflow}{An rflow object, e.g. as returned by \code{\link{rf}}.}
}
\value{
A logical value, whether the deletion was successful.
}
\description{
Forgets the computation for the current state.
}
