% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{forget}
\alias{forget}
\title{Forgets the computation for the current state.}
\usage{
forget(flow, state = "current")
}
\arguments{
\item{flow}{A flow object, e.g. as returned by \code{\link{flow_fn}}.}

\item{state}{A flow state. It can be either a valid state 
\code{index} (integer) or a valid state: \code{"current"}, \code{"all"}, 
\code{in_hash} or \code{out_hash} (string).}
}
\value{
A logical value, whether the deletion was successful.
}
\description{
Forgets the computation for the current state.
}
\examples{
fn <- function(x, y) { x + y + 10 }
make_flow_function <- make_flow_fn(fn)
rflow_function <- make_flow_function(2, 3) 
forget(rflow_function)

}
