% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nodes.R
\name{fun_signature}
\alias{fun_signature}
\title{Build Function Signature}
\usage{
fun_signature(func_name, func_args)
}
\arguments{
\item{func_name}{Character scalar giving function name}

\item{fun_args}{Data frame holding names and values of arguments}
}
\value{
Character scalar representing a function signature
}
\description{
Takes function name and argument names and values to build a signature
}

