% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{get_default_eddy}
\alias{get_default_eddy}
\title{Get the default eddy for the default environment.}
\usage{
get_default_eddy(cache_path = NULL, envir = get_default_env())
}
\arguments{
\item{cache_path}{A valid path of a directory to store the cache.
Use \code{NULL} (default) for no disk cache.}

\item{envir}{An environment where to find the default eddy.}
}
\value{
An R6Eddy object to be used when creating an rflow.
}
\description{
Get the default eddy for the default environment.
}
\seealso{
\code{\link{get_default_env}}
}
