% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{get_default_eddy}
\alias{get_default_eddy}
\title{Get the default, in memory, eddy for a given (or default) environment.}
\usage{
get_default_eddy(envir = get_default_env())
}
\arguments{
\item{envir}{An environment where to find the default eddy.}
}
\value{
An R6Eddy object to be used when creating an rflow.
}
\description{
Get the default, in memory, eddy for a given (or default) environment.
}
\seealso{
\code{\link{get_default_env}}
}
