% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{get_value}
\alias{get_value}
\title{Obtain value/object represented by a node}
\usage{
get_value(x, ...)
}
\arguments{
\item{x}{a node or node's id}

\item{rflow}{rflow object}
}
\value{
Value/object represented by given node.
}
\description{
Obtain value/object represented by a node
}
