% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODoutdata.R
\name{GW.nc}
\alias{GW.nc}
\title{MODFLOW results to NetCDF data set}
\usage{
GW.nc(dir, mfrt, file = paste0(mfrt, ".nc"), DIS = paste0(mfrt, ".dis"),
  HDS = paste0(mfrt, ".hds"), CBB = paste0(mfrt, ".cbb"),
  CBW = paste0(mfrt, ".cbw"), CRC = paste0(mfrt, ".crc"),
  CRV = paste0(mfrt, ".crv"), CBD = paste0(mfrt, ".cbd"),
  CBG = paste0(mfrt, ".cbg"), CS1 = paste0(mfrt, ".cs1"),
  BAS = paste0(mfrt, ".bas"), LPF = paste0(mfrt, ".lpf"),
  BCF = paste0(mfrt, ".bcf"), HNOFLO, HDRY,
  title = "untitled MODFLOW data set", author = "", datum = "",
  length.unit = "metre", time.unit = "day", xyt0 = c(0, 0, 0),
  start_date = "", updating = TRUE, .FillValue = 1e+30)
}
\arguments{
\item{dir}{directory in which reside the MODFLOW output and input files}

\item{mfrt}{non-extended filename to use as root for finding default
MODFLOW input and output file names}

\item{file}{filename for resulting NetCDF data set}

\item{HNOFLO}{value to interpret as inactive cells in head array,
automatically gleaned from BAS package file if not given; length-1
double}

\item{HDRY}{value to interpret as dry cell in head array, automatically
gleaned from LPF or BCF package file if not given; length-1 double}

\item{title}{of the model; length-1 character}

\item{author}{of the model; length-1 character}

\item{datum}{description of the reference datum for height/ head values,
for example "Ordnance Datum"; length-1 character}

\item{length.unit}{implied length units of the model; length-1 character}

\item{time.unit}{implied time units of the model; length-1 character}

\item{xyt0}{length-3 vector of the global x and y values at the bottom
left corner of the model and the absolute time value that the start of
the model refers to}

\item{start_date}{purely explanatory, enables the author to describe
what start_time refers to}

\item{updating}{get meta data (global attributes) from existing NetCDF
data set}
}
\value{
NULL.

Prints a summary of the final NetCDF data set using print.nc.
}
\description{
MODFLOW results to NetCDF data set
}
