% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_current}
\alias{is_current}
\title{Checks if the flow has a"current" state}
\usage{
is_current(flow)
}
\arguments{
\item{flow}{A flow object, e.g. as returned by \code{\link{flow_fn}}.}
}
\value{
A logical value, whether the current state is valid.
}
\description{
If there is no current state, e.g. right after \code{\link{make_flow_fn}},
  the flow is "not flowing", it is preventing downstream flows
  from being computed.
}
\examples{
fn <- function(x, y) { x + y + 8 }
flowed_fn <- flow_fn(1, 2, fn = fn) 
is_current_flow <- is_current(flowed_fn)

}
