% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_flow}
\alias{is_flow}
\title{Checks if the object is flow object or a flow element}
\usage{
is_flow(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value, whether \code{x} is a flow object.
}
\description{
Checks if the object is flow object or a flow element
}
\examples{
fn <- function(x, y) { x + y + 11 }
flow_function <- flow_fn(1, 2, fn = fn)
is_input_flow <- is_flow(flow_function)

}
