% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_flow}
\alias{is_flow}
\title{Is the object a flow object or a flow element?}
\usage{
is_flow(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value, whether \code{x} is a flow object.
}
\description{
Is the object a flow object or a flow element?
}
