% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_flow_fn}
\alias{is_flow_fn}
\title{Is the function an flow function (as returned by \code{make_flow_fn})?}
\usage{
is_flow_fn(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
A logical value, whether \code{fn} is a flow function.
}
\description{
Is the function an flow function (as returned by \code{make_flow_fn})?
}
