% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_valid}
\alias{is_valid}
\title{Is the rflow in a valid state?}
\usage{
is_valid(rflow)
}
\arguments{
\item{rflow}{An rflow object, e.g. as returned by \code{\link{rf}}.}
}
\value{
A logical value, whether the current state is valid.
}
\description{
It is used mostly for dignostic purposes. Invalid states also occur
  when the object is reconstructed from cache before any function call,
  i.e., as in the case of \code{\link{make_rf}}.
}
