% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{is_valid}
\alias{is_valid}
\title{Is the current state valid (stored in the cache)?}
\usage{
is_valid(flow, state = "current")
}
\arguments{
\item{flow}{An flow object, e.g. as returned by \code{\link{flow_call}}.}

\item{state}{An flow state. Only the \code{current} state is 
accepted for now.}
}
\value{
A logical value, whether the value can be obtained without
  triggering computation.
}
\description{
Is the current state valid (stored in the cache)?
}
