% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{load_state_of_nodes}
\alias{load_state_of_nodes}
\title{batch loading objects' definition/state from RDS files}
\usage{
load_state_of_nodes(path, recursive = FALSE, ignore.case = TRUE, ...)
}
\description{
batch loading objects' definition/state from RDS files
}
