% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{make_ns_sink}
\alias{make_ns_sink}
\title{Cache a function - TODO.}
\usage{
make_ns_sink(fn = to_ns, eddy = get_default_eddy())
}
\arguments{
\item{fn}{Function to be cached, the default works with environments and 
reactiveValues.}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the function.
}
\description{
Cache a function - TODO.
}
