% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{make_rflow}
\alias{make_rflow}
\title{Cache a function.}
\usage{
make_rflow(fn, hash_input_fn = NULL, split_output_fn = NULL,
  eddy = get_default_eddy())
}
\arguments{
\item{fn}{Function to be cached.}

\item{hash_input_fn}{Custom input hash function (e.g. a subset of RFlow output)}

\item{split_output_fn}{Function to split the output into a list,
in order to hash its elements separately.}

\item{eddy}{R6Eddy object were the data should be stored.}
}
\value{
The cached version of the inputted function.
}
\description{
Cache a function.
}
