% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{new_eddy}
\alias{new_eddy}
\title{Create a new eddy. Does not affect the current eddy}
\usage{
new_eddy(eddy_name, cache = default_cache(),
  flow_options = default_flow_options(), eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to allow retrieving later.}

\item{cache}{A cache object returned by one of the \code{cache} functions.}

\item{flow_options}{Options to store for future flow invocations. They
do not affect the \code{eddy}, they are only stored for ease of access.}

\item{eddy_env}{The environment in which the \code{eddy} is put (bound).}
}
\value{
An eddy object to be used for storing flows.
}
\description{
Create a new eddy. Does not affect the current eddy
}
\examples{
new_eddy("new_eddy_name")

}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_current_eddy}}, \code{\link{get_eddy}},
  \code{\link{set_current_eddy}}, \code{\link{use_eddy}}
}
\concept{eddy functions}
