% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy.R
\name{new_eddy}
\alias{new_eddy}
\alias{get_eddy}
\title{Create or retrieve an eddy.}
\usage{
new_eddy(cache_path = NULL, eddy_name = NULL, envir = get_default_env())

get_eddy(cache_path = NULL, eddy_name = NULL, envir = get_default_env())
}
\arguments{
\item{cache_path}{A valid path of a directory to store the cache.
Use \code{NULL} (default) for no disk cache.}

\item{eddy_name}{Name for the eddy folder.}

\item{envir}{An environment where to find the default eddy.}
}
\value{
An R6Eddy object to be used for storing rflows and their data.
}
\description{
Create or retrieve an eddy.
}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_default_eddy}}

Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_default_eddy}}
}
