% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_dag.R
\name{new_rflow}
\alias{new_rflow}
\title{Initialize a new DAG}
\usage{
new_rflow(path = NULL)
}
\arguments{
\item{path}{path to folder with nodes' definitions}
}
\value{
A new rflow object.
}
\description{
Initialize a new DAG
}
\examples{
RF <- new_rflow()
}
