% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{class}
\name{node}
\alias{node}
\title{basic building block of an rflow}
\format{\code{\link{R6Class}} object.}
\usage{
node("node1", desc = "This an example")
}
\value{
Object of \code{\link{R6Class}} with basic functionality for rflows.
}
\description{
basic building block of an rflow
}
\details{
Nodes may represent different types of data targets (R objects, files, database tables), or jobs.
Interconnected nodes together function as basic building elements of directed acyclic graph (DAG).

Nodes are implemented using R6 classes. There are various flavours of the basic node class (r_node, db_node, file_node) that have their specific requirements and features

Constructing nodes
\itemize{
\item from stored state
\item from lists (comming from TOML, YAML...)
}
}
\section{Fields}{

\describe{
\item{\code{id}}{character; unique id}

\item{\code{env}}{character; name of a environemnt (container or group of objects)}

\item{\code{name}}{character; object's name}

\item{\code{desc}}{character; description}

\item{\code{depends}}{character; vector of uppstream nodes (other nodes this node is depending on)}
}}

\section{Methods}{

\describe{
\item{\code{new(serveraddress)}}{This method is used to create object of this class with \code{serveraddress} as address of the server object is connecting to.}
}
}

\keyword{node}
