% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{node}
\alias{node}
\title{nodes can represent different type of data targets (R objects, files, database tables), or jobs...
interconnected nodes together function as basic building elements of directed acyclic graph (DAG)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
node
}
\description{
nodes are implemented using R6 classes with multiple descendants (r_node, db_node, file_node) that have their specific requirements and features
}
\details{
nodes are initialized
\itemize{
\item from stored state
\item from lists (comming from TOML, YAML...)
}
}
\keyword{datasets}
