% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{class}
\name{node}
\alias{node}
\title{node class}
\format{
\code{\link{R6Class}} object.
}
\usage{
node$new(id = "node")
}
\value{
Object of \code{\link{R6Class}} with basic functionality for rflows.
}
\description{
Node is a basic building block of an rflow DAG.
}
\details{
Nodes may represent different types of data targets (R objects, files, database tables), or jobs.
Interconnected nodes together function as basic building elements of directed acyclic graph (DAG).

Nodes are implemented using R6 classes. There are various flavors of the basic node class (r_node, db_node, file_node) that have their specific requirements and features

Constructing nodes
\itemize{
\item from stored state
\item from lists (coming from TOML, YAML...)
}
}
\section{Methods}{

\describe{
\item{\code{$make()} or \code{make(node)}}{This method is used to build/make targets. It recursively solves dependencies by building all targets for nodes declared or detected as a dependency.}
\item{\code{$eval()}}{Compared to \code{make()}, \code{eval()} only runs the code/job associated with the node alone. It does not try to ensure all the dependencies are ready. }
\item{\code{$process()}}{\code{process()} is usually used internally by \code{eval()} and should not be called directly}
\item{\code{$value}}{Returns target's value. Targets such as database tables or csv files are automatically convert into R values.}
\item{\code{$exists()}}{Checks whether the target exists or not.}
\item{\code{$time_to_eval()}}{Returns a record(s) of time to evaluate.}
}
}

\keyword{node}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character; unique id}

\item{\code{env}}{character; name of a environment (container or group of objects)}

\item{\code{name}}{character; object's name}

\item{\code{desc}}{character; description}

\item{\code{depends}}{character; vector of upstream nodes (other nodes this node is depending on)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vis_params_process}{\code{node$vis_params_process()}}
\item \href{#method-set_id}{\code{node$set_id()}}
\item \href{#method-set_persistence}{\code{node$set_persistence()}}
\item \href{#method-print}{\code{node$print()}}
\item \href{#method-label}{\code{node$label()}}
\item \href{#method-title}{\code{node$title()}}
\item \href{#method-new}{\code{node$new()}}
\item \href{#method-store_state}{\code{node$store_state()}}
\item \href{#method-update_definition}{\code{node$update_definition()}}
\item \href{#method-set_upstream}{\code{node$set_upstream()}}
\item \href{#method-set_downstream}{\code{node$set_downstream()}}
\item \href{#method-connect_to}{\code{node$connect_to()}}
\item \href{#method-connect}{\code{node$connect()}}
\item \href{#method-get}{\code{node$get()}}
\item \href{#method-remove}{\code{node$remove()}}
\item \href{#method-check_trigger_condition}{\code{node$check_trigger_condition()}}
\item \href{#method-check_triggers}{\code{node$check_triggers()}}
\item \href{#method-reset_triggers}{\code{node$reset_triggers()}}
\item \href{#method-eval}{\code{node$eval()}}
\item \href{#method-process}{\code{node$process()}}
\item \href{#method-changed}{\code{node$changed()}}
\item \href{#method-make}{\code{node$make()}}
\item \href{#method-validate}{\code{node$validate()}}
\item \href{#method-time_to_eval}{\code{node$time_to_eval()}}
\item \href{#method-clone}{\code{node$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vis_params_process"></a>}}
\if{latex}{\out{\hypertarget{method-vis_params_process}{}}}
\subsection{Method \code{vis_params_process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$vis_params_process(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$set_id(id, name, env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_persistence"></a>}}
\if{latex}{\out{\hypertarget{method-set_persistence}{}}}
\subsection{Method \code{set_persistence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$set_persistence(persistence)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\if{latex}{\out{\hypertarget{method-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-title"></a>}}
\if{latex}{\out{\hypertarget{method-title}{}}}
\subsection{Method \code{title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$new(
  id = NULL,
  name = NULL,
  env = NULL,
  desc = NULL,
  tags = NULL,
  depends = NULL,
  trigger_defchange = NULL,
  trigger_condition = NULL,
  definition_hash = NULL,
  persistence = list(enabled = FALSE),
  .last_evaluated = NULL,
  .last_changed = NULL,
  .trigger_defchange = NULL,
  .time_to_eval = NULL,
  vis_params = NULL,
  validators = NULL,
  logging = FALSE,
  loggers = NULL,
  store = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store_state"></a>}}
\if{latex}{\out{\hypertarget{method-store_state}{}}}
\subsection{Method \code{store_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$store_state(public_fields = NULL, private_fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_definition"></a>}}
\if{latex}{\out{\hypertarget{method-update_definition}{}}}
\subsection{Method \code{update_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$update_definition(
  id = NULL,
  type = NULL,
  desc = NULL,
  tags = NULL,
  depends = NULL,
  definition_hash = NULL,
  trigger_condition = NULL,
  vis_params = NULL,
  validators = NULL,
  ...,
  store = TRUE,
  logging = NULL,
  loggers = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_upstream"></a>}}
\if{latex}{\out{\hypertarget{method-set_upstream}{}}}
\subsection{Method \code{set_upstream()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$set_upstream(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_downstream"></a>}}
\if{latex}{\out{\hypertarget{method-set_downstream}{}}}
\subsection{Method \code{set_downstream()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$set_downstream(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to}{}}}
\subsection{Method \code{connect_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$connect_to(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$connect(verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$remove()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_trigger_condition"></a>}}
\if{latex}{\out{\hypertarget{method-check_trigger_condition}{}}}
\subsection{Method \code{check_trigger_condition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$check_trigger_condition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_triggers"></a>}}
\if{latex}{\out{\hypertarget{method-check_triggers}{}}}
\subsection{Method \code{check_triggers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$check_triggers(verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_triggers"></a>}}
\if{latex}{\out{\hypertarget{method-reset_triggers}{}}}
\subsection{Method \code{reset_triggers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$reset_triggers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$eval(verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$process(verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changed"></a>}}
\if{latex}{\out{\hypertarget{method-changed}{}}}
\subsection{Method \code{changed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$changed(verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make"></a>}}
\if{latex}{\out{\hypertarget{method-make}{}}}
\subsection{Method \code{make()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$make(force = FALSE, verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$validate(verbose = TRUE, verbose_prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time_to_eval"></a>}}
\if{latex}{\out{\hypertarget{method-time_to_eval}{}}}
\subsection{Method \code{time_to_eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$time_to_eval(what = getOption("RFLOW_TIME_TO_EVAL", "LAST"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{node$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
