% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rflow}
\alias{plot.rflow}
\title{visualize Rflow DAGs using visNetwork}
\usage{
\method{plot}{rflow}(
  x,
  tags = NULL,
  includeIsolated = TRUE,
  direction = "LR",
  tagsMatchLogic = "all",
  ...
)
}
\arguments{
\item{x}{an rflow objects}

\item{tags}{vector of tags for filtering nodes}

\item{includeIsolated}{logical; Switch to filter isolated/lonely nodes.}

\item{direction}{see visNetwork docs on hierarchical graphs}

\item{tagsMatchLogic}{one of: any/all/none}

\item{...}{args passed to visNetwork}
}
\description{
visualize Rflow DAGs using visNetwork
}
