% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_definition.R
\name{process_obj_defs}
\alias{process_obj_defs}
\title{Check consistency of identificators and fill missings}
\usage{
process_obj_defs(obj_defs)
}
\description{
Check consistency of identificators and fill missings
}
