% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEL.R
\name{read.WEL}
\alias{read.WEL}
\title{Read well (WEL) package file}
\usage{
read.WEL(filename, nSP, MF2k = TRUE)
}
\arguments{
\item{filename}{character string;
WEL package file}

\item{nSP}{integer \code{[1]} or DIS.MFpackage object;
number of stress periods or corresponding DIS package from which this
 can be read}

\item{MF2k}{logical \code{[1]};
is this a MODFLOW 2000+ WEL package?}
}
\value{
list \code{[2]} of class WEL.MFpackage, with elements:\cr
\code{$header}:\cr
  \code{..$MXACTW} the maximum number of active wells in a stress period\cr
  \code{..$IWELCB} unit number to which well flux array is saved (if
   \eqn{> 0})\cr
\code{$spheaders}:\cr
headers for each stress period, a data.frame with columns:\cr
  \code{..$ITMP} (int): number of active wells in stress period, or (if
   \eqn{< 0})\cr
  indicates to read from previous stress period\cr
  \code{..$NP} (int): number of parameters to read for this stress period
   (always 0 if \code{MF2k = FALSE})\cr
\code{$read}: logical vector of length nSP indicating for which stress periods
 well information is to be read\cr
\code{$data}:\cr
a data.table with the following columns\cr
  \code{..$sp} (int): stress period (key)\cr
  \code{..$C} (int): column\cr
  \code{..$R} (int): row\cr
  \code{..$L} (int): layer\cr
  \code{..$Q} (num): flux (negative indicates abstraction, positive indicates
   injection)\cr
}
\description{
Reads information from a MODFLOW WEL package file.  The WEL package
 assigns fixed sink/ source fluxes to individual cells and as such is
 particularly appropriate for representing wells of known abstraction or
 injection rate.
}
\examples{
fnms <- system.file(c("rflow_mf_demo.dis",
                      "rflow_mf_demo.wel"), package = "Rflow")

# get model information from DIS package file
dis <- read.DIS(fnms[1L])

wel <- read.WEL(fnms[2L], dis)
# or, if you already know how many stress periods there are
wel <- read.WEL(fnms[2L], 15L)

class(wel)
str(wel)

}

