% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODoutdata.R
\name{readHDS.arr}
\alias{readHDS.arr}
\title{Read Head Save array}
\usage{
readHDS.arr(file, conc = FALSE, time.only = FALSE, time.bn = 4L,
  hd.bn = 4L, show.help = TRUE, nf.to.NA = FALSE, nf.val = 999,
  CRs = list("all", "all"), lays = "all", sp_ts = "all")
}
\arguments{
\item{file}{character;
file name of head save array}

\item{conc}{logical;
is the file in fact an unformatted concentration file (MT3D output)?}

\item{time.only}{logical;
only read the model time information and discard the head data}

\item{time.bn}{integer, 4 or 8;
how many bytes do the time records occupy (generally 4)?}

\item{hd.bn}{integer, 4 or 8;
how many bytes to the head records occupy (generally 4)?}

\item{show.help}{logical;
show what is being read and print message describing output?}

\item{nf.to.NA}{logical;
convert no flow cells to NA}

\item{nf.val}{logical;
if nf.to.NA, what head value indicates a no flow cell?}

\item{CRs}{length-2 list;
elements are either "all" (default) or vectors of column/ row indices
which are requested}

\item{lays}{either "all" (default) or integer vector of requested layers}

\item{sp_ts}{either "all", a length-2 integer vector or a 2-column matrix;
which stress period-time step references to return}
}
\value{
a list with one or two elements:\cr
  \code{$Head}: num \code{[NCOL, NROW, NLAY, NTS]}; head values, only if
   \code{time.only = FALSE}
  \code{$time}: num \code{[NTS]}; time at end of each time step,
   relative to model start
}
\description{
Read Head Save array
}

