% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{rflow_end}
\alias{rflow_end}
\title{End RFlow}
\usage{
rflow_end(con)
}
\value{
NULL returned invisibly
}
\description{
Shuts down node-red environment, communication server and
  cleans up connection
}

