% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{rflow_send}
\alias{rflow_send}
\title{JSON Writer}
\usage{
rflow_send(request)
}
\arguments{
\item{request}{Character scalar holding JSON representation of the request}
}
\value{
Send request to the app and NULL invisibly
}
\description{
Send a JSON Representation of User-facing Functions to
  the NodeRed App
}

