% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nodes.R
\name{rflow_start}
\alias{rflow_start}
\title{Start RFlow}
\usage{
rflow_start(viewer = TRUE, tcp_port = 1338L)
}
\arguments{
\item{viewer}{Logical scalar determining whether to open the node-red environment
in the Rstudio viewer or a browser window}

\item{tcp_port}{Character scalar giving communication server port}
}
\value{
Opens Nodered app in the viewer or browser, and returns NULL
}
\description{
Takes app and communication ports to start
}

