% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{set_current_eddy}
\alias{set_current_eddy}
\title{Set the current eddy to be used in future flow calls.}
\usage{
set_current_eddy(eddy_name, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to to be set as current.}

\item{eddy_env}{The environment in which the eddy is put (bound).}
}
\value{
An eddy object to be used for storing flows.
}
\description{
Set the current eddy to be used in future flow calls.
}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_current_eddy}}, \code{\link{get_eddy}},
  \code{\link{new_eddy}}, \code{\link{use_eddy}}
}
\concept{eddy functions}
