% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink-ns.R
\name{to_ns}
\alias{to_ns}
\title{Assigns a value to a variable in a name space.}
\usage{
to_ns(x, var_name, ns)
}
\arguments{
\item{x}{Value to assign.}

\item{var_name}{The name (as string) of the variable.}

\item{ns}{The name space, either an \code{environment} or a 
\code{Shiny::reactiveValues} object.}
}
\value{
The initial value, \code{x}
}
\description{
Assigns a value to a variable in a name space.
}
