% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{trigger_manual}
\alias{trigger_manual}
\title{Obtain value/object represented by a node}
\usage{
trigger_manual(x, ...)
}
\arguments{
\item{x}{a node or node's id}

\item{rflow}{rflow object}
}
\value{
Value/object represented by given node.
}
\description{
Obtain value/object represented by a node
}
