% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflow_management.R
\name{update_node}
\alias{update_node}
\title{update object (call whenever definition in config file could have potentially changed)}
\usage{
update_node(obj, rflow, definition_hash = digest::digest(obj),
  verbose = FALSE)
}
\arguments{
\item{obj}{list with definition of node}

\item{rflow}{an rflow object}

\item{verbose}{logical; print verbose output?}
}
\description{
update object (call whenever definition in config file could have potentially changed)
}
