% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-functions.R
\name{use_eddy}
\alias{use_eddy}
\title{Convenience function that creates a new eddy or re-uses it if present}
\usage{
use_eddy(eddy_name, cache = default_cache(),
  flow_options = default_flow_options(), reuse_if_present = TRUE,
  set_current = TRUE, eddy_env = default_eddy_env())
}
\arguments{
\item{eddy_name}{Unique name for the eddy to allow retrieving later.}

\item{cache}{A cache object returned by one of the \code{cache} functions.}

\item{flow_options}{Options to store for future flow invocations. They
do not affect the \code{eddy}, they are only stored for ease of access.}

\item{reuse_if_present}{Logical, whether or not to reuse an eddy with 
the same name, if already present, even if the \code{flow_options}
 are different.}

\item{set_current}{Logical, whether to make this eddy current.}

\item{eddy_env}{The environment in which the eddy is put (bound).}
}
\value{
An eddy object to be used for storing flows.
}
\description{
Convenience function that creates a new eddy or re-uses it if present
}
\examples{
use_eddy("use_eddy")

}
\seealso{
Other eddy functions: \code{\link{delete_eddy}},
  \code{\link{get_current_eddy}}, \code{\link{get_eddy}},
  \code{\link{new_eddy}}, \code{\link{set_current_eddy}}
}
\concept{eddy functions}
