% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visRflow2}
\alias{visRflow2}
\title{visualize workflow DAG}
\usage{
visRflow2(rflow, tags = NULL, includeIsolated = TRUE,
  direction = "UD", orderBy = "id", ...)
}
\arguments{
\item{...}{}
}
\description{
visualize workflow DAG
}
