% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visRflow2}
\alias{visRflow2}
\title{visualize workflow DAG}
\usage{
visRflow2(
  rflow,
  tags = NULL,
  includeIsolated = TRUE,
  direction = "UD",
  orderBy = "id",
  ...
)
}
\arguments{
\item{rflow}{an rflow object}

\item{tags}{character; vector of tags or regular expression for filtering nodes by tags}

\item{includeIsolated}{logical; inlude disconnected nodes?}

\item{direction}{orientation of flow ("UD" by default)}

\item{...}{}
}
\value{
A visNetwork plot.
}
\description{
visualize workflow DAG
}
