% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIS.R
\name{write.DIS}
\alias{write.DIS}
\title{Write a discretisation (DIS) package file}
\usage{
write.DIS(DIS, filename, title)
}
\arguments{
\item{DIS}{object of class DIS.MFpackage, as would be read by
\code{\link{read.DIS}}}

\item{filename}{character string;
file name to write to, ideally ending in \code{".dis"}}

\item{title}{character string optional title to put at the start of the
package file}
}
\value{
\code{NULL}
}
\description{
Writes information from a DIS.MFpackage list object to a MODFLOW-readable
 DIS package file.
}
\examples{
# read DIS package
fnm <- system.file("rflow_mf_demo.dis", package = "Rflow")
dis <- read.DIS(fnm)

# meaningful modification e.g. add a stress period
dis$extent["NPER"] <- dis$extent["NPER"] + 1L
dis$sps <- rbind(dis$sps,
                 list(1000, 10L, 1.2, TRUE))

# write modified dis package to file
write.DIS(dis, "RFLOW_EXAMPLE.dis", "example: added stress period")

}
