% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFB.R
\name{write.HFB}
\alias{write.HFB}
\title{Write a hydraulic flow barrier (HFB) package file}
\usage{
write.HFB(HFB, filename)
}
\arguments{
\item{HFB}{HFB.MFpackage object, as would be returned by \code{\link{read.HFB}}}

\item{filename}{character string}
}
\description{
Writes information from a HFB.MFpackage list object to a MODFLOW-readable
 HFB package file.
}
